Use weblate for localization (#1777)
* Update language json format for weblate

* Updated language patch code and test

* Updated readme for weblate

* Cleaned up language files
