+++ Nitrox.Assets.Subnautica/LanguageFiles/Croatian.json
@@ -1,44 +1,38 @@
 {
-"version": 3,
-// Main menu
-"Nitrox_Multiplayer": "Multiplayer",
-"Nitrox_AddServer": "Dodaj Server",
-"Nitrox_ConnectTo": "Spajanje na",
-"Nitrox_AddServerName": "Naziv:",
-"Nitrox_AddServerHost": "Host:",
-"Nitrox_AddServerPort": "Port:",
-"Nitrox_AddServerAdd": "Dodaj",
-"Nitrox_JoinServerPasswordHeader": "Obavezna server lozinka",
-"Nitrox_JoinServerIpAddress": "Server IP",
-"Nitrox_JoinServerPassword": "Lozinka:",
-"Nitrox_JoinServerFailed": "Nije se moguće spojiti sa serverom",
-"Nitrox_Cancel": "Odustani",
-"Nitrox_Join": "Pridruži se",
-"Nitrox_PlayerColor": "Boja igrača",
-"Nitrox_EnterName": "Ime igrača",
-"Nitrox_SubmitPassword": "Potvrdi",
-"Nitrox_MultiplayerLoaded": "Multiplayer klijent učitan...",
-"Nitrox_StreamerSettings": "Streamer mode",
-"Nitrox_SilenceChat": "Utišaj razgovor (chat)",
-"Nitrox_HideIp": "Sakrij IPs",
-// In game menu
-"Nitrox_LoadingScreenWarn": "Očekujte probleme koje mogu skršiti vaš gameplay",
-"Nitrox_WorldSettling": "Awaiting World Settling",
-"Nitrox_JoiningSession": "Pridruživanje multiplayer sessionu",
-"Nitrox_SyncingWorld": "Sinkronizacija multiplayer svijeta",
-"Nitrox_LostConnection": "Izgubljena veza sa serverom",
-"Nitrox_PlayerKicked": "Kickan si sa servera",
-// In game logging
-"Nitrox_UnableToConnect": "Nije moguće kontaktirati server na:",
-"Nitrox_StartServer": "Prvo pokrenite svoj server da biste se pridružili u igru koju sami hostirate",
-"Nitrox_FirewallInterfering": "Čini se da postavke vašeg firewalla ometaju",
-"Nitrox_InvalidUserName": "Molimo vas da unesete valjano ime !\n\n Ne smije sadržavati razmak ili sumnjive znakove\n Dozvoljeni znakovi : A-Z a-z 0-9 _ . -\nDužina : [3, 25]",
-"Nitrox_RequestingSessionPolicy": "Zahtjevanje informacija o pravilima sesije...",
-"Nitrox_RejectedSessionPolicy": "Rezervacija odbijena...",
-"Nitrox_DisconnectedSession": "Prekinuta veza sa serverom",
-"Nitrox_Waiting": "Čekanje u redu za pridruživanje...",
-"Nitrox_WaitingPassword": "Čeka se unos lozinke servera...",
-"Nitrox_WaitingUserInput": "Čeka se korisnika na unos...",
-"Nitrox_LaunchGame": "Pokretanje igre...",
-"Nitrox_SilencedChatNotif": "Razgovor je utišan"
+    "Nitrox_Multiplayer": "Multiplayer",
+    "Nitrox_AddServer": "Dodaj Server",
+    "Nitrox_ConnectTo": "Spajanje na",
+    "Nitrox_AddServerName": "Naziv:",
+    "Nitrox_AddServerHost": "Host:",
+    "Nitrox_AddServerPort": "Port:",
+    "Nitrox_AddServerAdd": "Dodaj",
+    "Nitrox_JoinServerPasswordHeader": "Obavezna server lozinka",
+    "Nitrox_JoinServerIpAddress": "Server IP",
+    "Nitrox_JoinServerPassword": "Lozinka:",
+    "Nitrox_JoinServerFailed": "Nije se moguće spojiti sa serverom",
+    "Nitrox_Cancel": "Odustani",
+    "Nitrox_Join": "Pridruži se",
+    "Nitrox_PlayerColor": "Boja igrača",
+    "Nitrox_EnterName": "Ime igrača",
+    "Nitrox_SubmitPassword": "Potvrdi",
+    "Nitrox_MultiplayerLoaded": "Multiplayer klijent učitan…",
+    "Nitrox_SilenceChat": "Utišaj razgovor (chat)",
+    "Nitrox_HideIp": "Sakrij IPs",
+    "Nitrox_LoadingScreenWarn": "Očekujte probleme koje mogu skršiti vaš gameplay",
+    "Nitrox_JoiningSession": "Pridruživanje multiplayer sessionu",
+    "Nitrox_SyncingWorld": "Sinkronizacija multiplayer svijeta",
+    "Nitrox_LostConnection": "Izgubljena veza sa serverom",
+    "Nitrox_PlayerKicked": "Kickan si sa servera",
+    "Nitrox_UnableToConnect": "Nije moguće kontaktirati server na:",
+    "Nitrox_StartServer": "Prvo pokrenite svoj server da biste se pridružili u igru koju sami hostirate",
+    "Nitrox_FirewallInterfering": "Čini se da postavke vašeg firewalla ometaju",
+    "Nitrox_InvalidUserName": "Molimo vas da unesete valjano ime !\n\n Ne smije sadržavati razmak ili sumnjive znakove\n Dozvoljeni znakovi : A-Z a-z 0-9 _ . -\nDužina : [3, 25]",
+    "Nitrox_RequestingSessionPolicy": "Zahtjevanje informacija o pravilima sesije…",
+    "Nitrox_RejectedSessionPolicy": "Rezervacija odbijena…",
+    "Nitrox_DisconnectedSession": "Prekinuta veza sa serverom",
+    "Nitrox_Waiting": "Čekanje u redu za pridruživanje…",
+    "Nitrox_WaitingPassword": "Čeka se unos lozinke servera…",
+    "Nitrox_WaitingUserInput": "Čeka se korisnika na unos…",
+    "Nitrox_LaunchGame": "Pokretanje igre…",
+    "Nitrox_SilencedChatNotif": "Razgovor je utišan"
 }
