+++ Nitrox.Assets.Subnautica/LanguageFiles/Czech.json
@@ -1,44 +1,40 @@
 {
-"version": 3,
-// Main menu
-"Nitrox_Multiplayer": "Multiplayer",
-"Nitrox_AddServer": "Přidat Server",
-"Nitrox_ConnectTo": "Připojit se na",
-"Nitrox_AddServerName": "Jméno:",
-"Nitrox_AddServerHost": "Host:",
-"Nitrox_AddServerPort": "Port:",
-"Nitrox_AddServerAdd": "Přidat",
-"Nitrox_JoinServerPasswordHeader": "Server vyžaduje heslo",
-"Nitrox_JoinServerIpAddress": "IP Serveru",
-"Nitrox_JoinServerPassword": "Heslo:",
-"Nitrox_JoinServerFailed": "Nepodařilo se připojit k serveru",
-"Nitrox_Cancel": "Zrušit",
-"Nitrox_Join": "Připojit se",
-"Nitrox_PlayerColor": "Barva Hráče",
-"Nitrox_EnterName": "Jméno Hráče",
-"Nitrox_SubmitPassword": "Potvrdit",
-"Nitrox_MultiplayerLoaded": "Multiplayerový klient načten...",
-"Nitrox_StreamerSettings": "Režim streamování",
-"Nitrox_SilenceChat": "Tichý chat",
-"Nitrox_HideIp": "Skrýt IP adresu",
-// In game menu
-"Nitrox_LoadingScreenWarn": "Očekávejte herní chyby které mohou kazit hru",
-"Nitrox_WorldSettling": "Čekání na načtení světa",
-"Nitrox_JoiningSession": "Připojování k serveru",
-"Nitrox_SyncingWorld": "Synchronizování Multiplayerového Světa",
-"Nitrox_LostConnection": "Připojení k serveru ztraceno",
-"Nitrox_PlayerKicked": "Byl jsi vyhozen ze serveru",
-// In game logging
-"Nitrox_UnableToConnect": "Nepodařilo se kontaktovat vzdálený server na:",
-"Nitrox_StartServer": "Pro připojení na vlastní hostovaný svět je potřeba zapnout server",
-"Nitrox_FirewallInterfering": "Vypadá to že nastavení vaší brány firewall je nesprávné",
-"Nitrox_InvalidUserName": "Prosím vložte platné hráčské jméno !\n\n Nesmí obsahovat žádnou mezeru nebo pochybné znaky\n Povolené znaky : A-Z a-z 0-9 _ . -\nDélka : [3, 25]",
-"Nitrox_RequestingSessionPolicy": "Požadování informací o stávající relaci...",
-"Nitrox_RejectedSessionPolicy": "Rezervace odmítnuta...",
-"Nitrox_DisconnectedSession": "Odpojen ze serveru",
-"Nitrox_Waiting": "Čekání ve frontě připojení...",
-"Nitrox_WaitingPassword": "Čekání na zadání serverového hesla...",
-"Nitrox_WaitingUserInput": "Čekání na uživatelský vstup...",
-"Nitrox_LaunchGame": "Spouštení hry...",
-"Nitrox_SilencedChatNotif": "Chat je nyní tichý"
+    "Nitrox_Multiplayer": "Multiplayer",
+    "Nitrox_AddServer": "Přidat Server",
+    "Nitrox_ConnectTo": "Připojit se na",
+    "Nitrox_AddServerName": "Jméno:",
+    "Nitrox_AddServerHost": "Host:",
+    "Nitrox_AddServerPort": "Port:",
+    "Nitrox_AddServerAdd": "Přidat",
+    "Nitrox_JoinServerPasswordHeader": "Server vyžaduje heslo",
+    "Nitrox_JoinServerIpAddress": "IP Serveru",
+    "Nitrox_JoinServerPassword": "Heslo:",
+    "Nitrox_JoinServerFailed": "Nepodařilo se připojit k serveru",
+    "Nitrox_Cancel": "Zrušit",
+    "Nitrox_Join": "Připojit se",
+    "Nitrox_PlayerColor": "Barva Hráče",
+    "Nitrox_EnterName": "Jméno Hráče",
+    "Nitrox_SubmitPassword": "Potvrdit",
+    "Nitrox_MultiplayerLoaded": "Multiplayerový klient načten…",
+    "Nitrox_StreamerSettings": "Režim streamování",
+    "Nitrox_SilenceChat": "Tichý chat",
+    "Nitrox_HideIp": "Skrýt IP adresu",
+    "Nitrox_LoadingScreenWarn": "Očekávejte herní chyby které mohou kazit hru",
+    "Nitrox_WorldSettling": "Čekání na načtení světa",
+    "Nitrox_JoiningSession": "Připojování k serveru",
+    "Nitrox_SyncingWorld": "Synchronizování Multiplayerového Světa",
+    "Nitrox_LostConnection": "Připojení k serveru ztraceno",
+    "Nitrox_PlayerKicked": "Byl jsi vyhozen ze serveru",
+    "Nitrox_UnableToConnect": "Nepodařilo se kontaktovat vzdálený server na:",
+    "Nitrox_StartServer": "Pro připojení na vlastní hostovaný svět je potřeba zapnout server",
+    "Nitrox_FirewallInterfering": "Vypadá to že nastavení vaší brány firewall je nesprávné",
+    "Nitrox_InvalidUserName": "Prosím vložte platné hráčské jméno !\n\n Nesmí obsahovat žádnou mezeru nebo pochybné znaky\n Povolené znaky : A-Z a-z 0-9 _ . -\nDélka : [3, 25]",
+    "Nitrox_RequestingSessionPolicy": "Požadování informací o stávající relaci…",
+    "Nitrox_RejectedSessionPolicy": "Rezervace odmítnuta…",
+    "Nitrox_DisconnectedSession": "Odpojen ze serveru",
+    "Nitrox_Waiting": "Čekání ve frontě připojení…",
+    "Nitrox_WaitingPassword": "Čekání na zadání serverového hesla…",
+    "Nitrox_WaitingUserInput": "Čekání na uživatelský vstup…",
+    "Nitrox_LaunchGame": "Spouštení hry…",
+    "Nitrox_SilencedChatNotif": "Chat je nyní tichý"
 }
