+++ Nitrox.Assets.Subnautica/LanguageFiles/Danish.json
@@ -1,44 +1,40 @@
 {
-"version": 3,
-// Main menu
-"Nitrox_Multiplayer": "Multiplayer",
-"Nitrox_AddServer": "Tilføj server",
-"Nitrox_ConnectTo": "Forbind til",
-"Nitrox_AddServerName": "Navn:",
-"Nitrox_AddServerHost": "Host:",
-"Nitrox_AddServerPort": "Port:",
-"Nitrox_AddServerAdd": "Tilføj",
-"Nitrox_JoinServerPasswordHeader": "Server adgangskode påkrævet",
-"Nitrox_JoinServerIpAddress": "Server IP",
-"Nitrox_JoinServerPassword": "Adgangskode:",
-"Nitrox_JoinServerFailed": "Kunne ikke forbinde til serveren",
-"Nitrox_Cancel": "Annuller",
-"Nitrox_Join": "Deltag",
-"Nitrox_PlayerColor": "Spillerfarve",
-"Nitrox_EnterName": "Spillernavn",
-"Nitrox_SubmitPassword": "Forbind",
-"Nitrox_MultiplayerLoaded": "Multiplayer-klient indlæst...",
-"Nitrox_StreamerSettings": "Streamer-tilstand",
-"Nitrox_SilenceChat": "Slå chat fra",
-"Nitrox_HideIp": "Skjul IP-adresser",
-// In game menu
-"Nitrox_LoadingScreenWarn": "Forvent spil-ødelæggene fejl",
-"Nitrox_WorldSettling": "Afventer verdens afvikling",
-"Nitrox_JoiningSession": "Deltager i spil",
-"Nitrox_SyncingWorld": "Synkronisere multiplayer verden",
-"Nitrox_LostConnection": "Mistede forbindelse til spil Server",
-"Nitrox_PlayerKicked": "Du er blevet smidt ud fra serveren:",
-// In game logging
-"Nitrox_UnableToConnect": "Kan ikke kontakte den eksterne server på:",
-"Nitrox_StartServer": "Start din server først for at deltage i din selvhostede verden",
-"Nitrox_FirewallInterfering": "Det ser ud til, at dine firewallindstillinger forstyrrer",
-"Nitrox_InvalidUserName": "Indtast venligst et gyldigt navn!\n\n Det må ikke indeholde mellemrum eller tvivlsomme tegn\n Tilladte tegn : A-Z a-z 0-9 _ . -\nLængde : [3, 25]",
-"Nitrox_RequestingSessionPolicy": "Anmoder om sessionens politik...",
-"Nitrox_RejectedSessionPolicy": "Reservation afvist...",
-"Nitrox_DisconnectedSession": "Afbrudt fra server",
-"Nitrox_Waiting": "Venter i forbindelseskøen...",
-"Nitrox_WaitingPassword": "Venter på serveradgangskode...",
-"Nitrox_WaitingUserInput": "Venter på brugerens input...",
-"Nitrox_LaunchGame": "Starter spil...",
-"Nitrox_SilencedChatNotif": "Chat er nu slået fra"
+    "Nitrox_Multiplayer": "Multiplayer",
+    "Nitrox_AddServer": "Tilføj server",
+    "Nitrox_ConnectTo": "Forbind til",
+    "Nitrox_AddServerName": "Navn:",
+    "Nitrox_AddServerHost": "Host:",
+    "Nitrox_AddServerPort": "Port:",
+    "Nitrox_AddServerAdd": "Tilføj",
+    "Nitrox_JoinServerPasswordHeader": "Server adgangskode påkrævet",
+    "Nitrox_JoinServerIpAddress": "Server IP",
+    "Nitrox_JoinServerPassword": "Adgangskode:",
+    "Nitrox_JoinServerFailed": "Kunne ikke forbinde til serveren",
+    "Nitrox_Cancel": "Annuller",
+    "Nitrox_Join": "Deltag",
+    "Nitrox_PlayerColor": "Spillerfarve",
+    "Nitrox_EnterName": "Spillernavn",
+    "Nitrox_SubmitPassword": "Forbind",
+    "Nitrox_MultiplayerLoaded": "Multiplayer-klient indlæst…",
+    "Nitrox_StreamerSettings": "Streamer-tilstand",
+    "Nitrox_SilenceChat": "Slå chat fra",
+    "Nitrox_HideIp": "Skjul IP-adresser",
+    "Nitrox_LoadingScreenWarn": "Forvent spil-ødelæggene fejl",
+    "Nitrox_WorldSettling": "Afventer verdens afvikling",
+    "Nitrox_JoiningSession": "Deltager i spil",
+    "Nitrox_SyncingWorld": "Synkronisere multiplayer verden",
+    "Nitrox_LostConnection": "Mistede forbindelse til spil Server",
+    "Nitrox_PlayerKicked": "Du er blevet smidt ud fra serveren",
+    "Nitrox_UnableToConnect": "Kan ikke kontakte den eksterne server på:",
+    "Nitrox_StartServer": "Start din server først for at deltage i din selvhostede verden",
+    "Nitrox_FirewallInterfering": "Det ser ud til, at dine firewallindstillinger forstyrrer",
+    "Nitrox_InvalidUserName": "Indtast venligst et gyldigt navn!\n\n Det må ikke indeholde mellemrum eller tvivlsomme tegn\n Tilladte tegn : A-Z a-z 0-9 _ . -\nLængde : [3, 25]",
+    "Nitrox_RequestingSessionPolicy": "Anmoder om sessionens politik…",
+    "Nitrox_RejectedSessionPolicy": "Reservation afvist…",
+    "Nitrox_DisconnectedSession": "Afbrudt fra server",
+    "Nitrox_Waiting": "Venter i forbindelseskøen…",
+    "Nitrox_WaitingPassword": "Venter på serveradgangskode…",
+    "Nitrox_WaitingUserInput": "Venter på brugerens input…",
+    "Nitrox_LaunchGame": "Starter spil…",
+    "Nitrox_SilencedChatNotif": "Chat er nu slået fra"
 }
