+++ Nitrox.Assets.Subnautica/LanguageFiles/Dutch.json
@@ -1,44 +1,40 @@
 {
-"version": 3,
-// Main menu
-"Nitrox_Multiplayer": "Multiplayer",
-"Nitrox_AddServer": "Server toevoegen",
-"Nitrox_ConnectTo": "Verbinden met",
-"Nitrox_AddServerName": "Naam:",
-"Nitrox_AddServerHost": "Host:",
-"Nitrox_AddServerPort": "Port:",
-"Nitrox_AddServerAdd": "Toevoegen",
-"Nitrox_JoinServerPasswordHeader": "Server wachtwoord vereist",
-"Nitrox_JoinServerIpAddress": "Server IP",
-"Nitrox_JoinServerPassword": "Wachtwoord:",
-"Nitrox_JoinServerFailed": "Kan niet verbinden met server",
-"Nitrox_Cancel": "Annuleer",
-"Nitrox_Join": "Meedoen",
-"Nitrox_PlayerColor": "Speler Kleur",
-"Nitrox_EnterName": "Speler Naam",
-"Nitrox_SubmitPassword": "Verzenden",
-"Nitrox_MultiplayerLoaded": "Multiplayer client geladen...",
-"Nitrox_StreamerSettings": "Streamer modus",
-"Nitrox_SilenceChat": "Chat dempen",
-"Nitrox_HideIp": "IPs verschuilen",
-// In game menu
-"Nitrox_LoadingScreenWarn": "Verwacht spel-brekende bugs",
-"Nitrox_WorldSettling": "Wachtend op ruststand van spel wereld",
-"Nitrox_JoiningSession": "Meedoen aan multiplayer Sessie",
-"Nitrox_SyncingWorld": "Multiplayer Wereld Synchroniseren",
-"Nitrox_LostConnection": "Connectie met de server verloren",
-"Nitrox_PlayerKicked": "Je bent van de server getrapt",
-// In game logging
-"Nitrox_UnableToConnect": "Kan niet verbinden met de externe server op:",
-"Nitrox_StartServer": "Start je server eerst om mee toe doen met je eigen gehostte wereld",
-"Nitrox_FirewallInterfering": "Het lijkt erop dat je firewall instellingen het netwerk onderbreken",
-"Nitrox_InvalidUserName": "Voer een geldige spelernaam in !\n\n Het mag geen spatie of speciale karakters bevatten\n Toegestane karakters : A-Z a-z 0-9 _ . -\nLengte : [3, 25]",
-"Nitrox_RequestingSessionPolicy": "Beleid van de sessie aanvragen...",
-"Nitrox_RejectedSessionPolicy": "Sessie geweigerd...",
-"Nitrox_DisconnectedSession": "U heeft de server verlaten",
-"Nitrox_Waiting": "Wachtend om mee te doen...",
-"Nitrox_WaitingPassword": "Wachtend op invoer van de server...",
-"Nitrox_WaitingUserInput": "Wachtend op invoer van de gebruiker...",
-"Nitrox_LaunchGame": "Spel opstarten...",
-"Nitrox_SilencedChatNotif": "Chat is gedempt"
+    "Nitrox_Multiplayer": "Multiplayer",
+    "Nitrox_AddServer": "Server toevoegen",
+    "Nitrox_ConnectTo": "Verbinden met",
+    "Nitrox_AddServerName": "Naam:",
+    "Nitrox_AddServerHost": "Host:",
+    "Nitrox_AddServerPort": "Port:",
+    "Nitrox_AddServerAdd": "Toevoegen",
+    "Nitrox_JoinServerPasswordHeader": "Server wachtwoord vereist",
+    "Nitrox_JoinServerIpAddress": "Server IP",
+    "Nitrox_JoinServerPassword": "Wachtwoord:",
+    "Nitrox_JoinServerFailed": "Kan niet verbinden met server",
+    "Nitrox_Cancel": "Annuleer",
+    "Nitrox_Join": "Meedoen",
+    "Nitrox_PlayerColor": "Speler Kleur",
+    "Nitrox_EnterName": "Speler Naam",
+    "Nitrox_SubmitPassword": "Verzenden",
+    "Nitrox_MultiplayerLoaded": "Multiplayer client geladen…",
+    "Nitrox_StreamerSettings": "Streamer modus",
+    "Nitrox_SilenceChat": "Chat dempen",
+    "Nitrox_HideIp": "IPs verschuilen",
+    "Nitrox_LoadingScreenWarn": "Verwacht spel-brekende bugs",
+    "Nitrox_WorldSettling": "Wachtend op ruststand van spel wereld",
+    "Nitrox_JoiningSession": "Meedoen aan multiplayer Sessie",
+    "Nitrox_SyncingWorld": "Multiplayer Wereld Synchroniseren",
+    "Nitrox_LostConnection": "Connectie met de server verloren",
+    "Nitrox_PlayerKicked": "Je bent van de server getrapt",
+    "Nitrox_UnableToConnect": "Kan niet verbinden met de externe server op:",
+    "Nitrox_StartServer": "Start je server eerst om mee toe doen met je eigen gehostte wereld",
+    "Nitrox_FirewallInterfering": "Het lijkt erop dat je firewall instellingen het netwerk onderbreken",
+    "Nitrox_InvalidUserName": "Voer een geldige spelernaam in !\n\n Het mag geen spatie of speciale karakters bevatten\n Toegestane karakters : A-Z a-z 0-9 _ . -\nLengte : [3, 25]",
+    "Nitrox_RequestingSessionPolicy": "Beleid van de sessie aanvragen…",
+    "Nitrox_RejectedSessionPolicy": "Sessie geweigerd…",
+    "Nitrox_DisconnectedSession": "U heeft de server verlaten",
+    "Nitrox_Waiting": "Wachtend om mee te doen…",
+    "Nitrox_WaitingPassword": "Wachtend op invoer van de server…",
+    "Nitrox_WaitingUserInput": "Wachtend op invoer van de gebruiker…",
+    "Nitrox_LaunchGame": "Spel opstarten…",
+    "Nitrox_SilencedChatNotif": "Chat is gedempt"
 }
