+++ Nitrox.Assets.Subnautica/LanguageFiles/English.json
@@ -1,51 +1,47 @@
 {
-"version": 5,
-// Main menu
-"Nitrox_Multiplayer": "Multiplayer",
-"Nitrox_AddServer": "Add Server",
-"Nitrox_ConnectTo": "Connect to",
-"Nitrox_AddServerName": "Name:",
-"Nitrox_AddServerHost": "Host:",
-"Nitrox_AddServerPort": "Port:",
-"Nitrox_AddServerAdd": "Add",
-"Nitrox_JoinServerPasswordHeader": "Server password required",
-"Nitrox_JoinServerIpAddress": "Server IP",
-"Nitrox_JoinServerPassword": "Password:",
-"Nitrox_JoinServerFailed": "Unable to join server",
-"Nitrox_Cancel": "Cancel",
-"Nitrox_Join": "Join",
-"Nitrox_PlayerColor": "Player Color",
-"Nitrox_EnterName": "Player Name",
-"Nitrox_SubmitPassword": "Submit",
-"Nitrox_MultiplayerLoaded": "Multiplayer client loaded...",
-"Nitrox_StreamerSettings": "Streamer mode",
-"Nitrox_SilenceChat": "Mute chat",
-"Nitrox_HideIp": "Hide IP addresses",
-// In game menu
-"Nitrox_LoadingScreenWarn": "Expect game-breaking bugs",
-"Nitrox_WorldSettling": "Awaiting World Settling",
-"Nitrox_JoiningSession": "Joining Multiplayer Session",
-"Nitrox_SyncingWorld": "Syncing Multiplayer World",
-"Nitrox_LostConnection": "Lost connection to game Server",
-"Nitrox_PlayerKicked": "You've been kicked from the server",
-"Nitrox_LostConnectionWarn": "No longer connected to the server",
-"Nitrox_ServerStopped": "Server stopped",
-// In game logging
-"Nitrox_UnableToConnect": "Unable to contact the remote server at:",
-"Nitrox_StartServer": "Start your server first to join your self-hosted world",
-"Nitrox_FirewallInterfering": "Seems like your firewall settings are interfering",
-"Nitrox_InvalidUserName": "Please enter a valid player name !\n\n It must not contain any space or doubtful characters\n Allowed characters : A-Z a-z 0-9 _ . -\nLength : [3, 25]",
-"Nitrox_RequestingSessionPolicy": "Requesting session policy information...",
-"Nitrox_RejectedSessionPolicy": "Reservation rejected...",
-"Nitrox_DisconnectedSession": "Disconnected from server",
-"Nitrox_Waiting": "Waiting in join queue...",
-"Nitrox_WaitingPassword": "Waiting for server password input...",
-"Nitrox_WaitingUserInput": "Waiting for user input...",
-"Nitrox_LaunchGame": "Launching game...",
-"Nitrox_SilencedChatNotif": "Chat is now muted",
-"Nitrox_CommandNotAvailable": "This command is not available in Nitrox",
-"Nitrox_DiscordAccept": "Accept",
-"Nitrox_DiscordDecline": "Decline",
-"Nitrox_DiscordPressToFocus": "Press to unlock",
-"Nitrox_DiscordRequestText": "Requested to join your session"
+    "Nitrox_Multiplayer": "Multiplayer",
+    "Nitrox_AddServer": "Add Server",
+    "Nitrox_ConnectTo": "Connect to",
+    "Nitrox_AddServerName": "Name:",
+    "Nitrox_AddServerHost": "Host:",
+    "Nitrox_AddServerPort": "Port:",
+    "Nitrox_AddServerAdd": "Add",
+    "Nitrox_JoinServerPasswordHeader": "Server password required",
+    "Nitrox_JoinServerIpAddress": "Server IP",
+    "Nitrox_JoinServerPassword": "Password:",
+    "Nitrox_JoinServerFailed": "Unable to join server",
+    "Nitrox_Cancel": "Cancel",
+    "Nitrox_Join": "Join",
+    "Nitrox_PlayerColor": "Player Color",
+    "Nitrox_EnterName": "Player Name",
+    "Nitrox_SubmitPassword": "Submit",
+    "Nitrox_MultiplayerLoaded": "Multiplayer client loaded…",
+    "Nitrox_StreamerSettings": "Streamer mode",
+    "Nitrox_SilenceChat": "Mute chat",
+    "Nitrox_HideIp": "Hide IP addresses",
+    "Nitrox_LoadingScreenWarn": "Expect game-breaking bugs",
+    "Nitrox_WorldSettling": "Awaiting World Settling",
+    "Nitrox_JoiningSession": "Joining Multiplayer Session",
+    "Nitrox_SyncingWorld": "Syncing Multiplayer World",
+    "Nitrox_LostConnection": "Lost connection to game Server",
+    "Nitrox_PlayerKicked": "You've been kicked from the server",
+    "Nitrox_LostConnectionWarn": "No longer connected to the server",
+    "Nitrox_ServerStopped": "Server stopped",
+    "Nitrox_UnableToConnect": "Unable to contact the remote server at:",
+    "Nitrox_StartServer": "Start your server first to join your self-hosted world",
+    "Nitrox_FirewallInterfering": "Seems like your firewall settings are interfering",
+    "Nitrox_InvalidUserName": "Please enter a valid player name !\n\n It must not contain any space or doubtful characters\n Allowed characters : A-Z a-z 0-9 _ . -\nLength : [3, 25]",
+    "Nitrox_RequestingSessionPolicy": "Requesting session policy information…",
+    "Nitrox_RejectedSessionPolicy": "Reservation rejected…",
+    "Nitrox_DisconnectedSession": "Disconnected from server",
+    "Nitrox_Waiting": "Waiting in join queue…",
+    "Nitrox_WaitingPassword": "Waiting for server password input…",
+    "Nitrox_WaitingUserInput": "Waiting for user input…",
+    "Nitrox_LaunchGame": "Launching game…",
+    "Nitrox_SilencedChatNotif": "Chat is now muted",
+    "Nitrox_CommandNotAvailable": "This command is not available in Nitrox",
+    "Nitrox_DiscordAccept": "Accept",
+    "Nitrox_DiscordDecline": "Decline",
+    "Nitrox_DiscordPressToFocus": "Press to unlock",
+    "Nitrox_DiscordRequestText": "Requested to join your session"
 }
