+++ Nitrox.Assets.Subnautica/LanguageFiles/Finnish.json
@@ -1,44 +1,40 @@
 {
-"version": 3,
-// Main menu
-"Nitrox_Multiplayer": "Moninpeli",
-"Nitrox_AddServer": "Lisää palvelin",
-"Nitrox_ConnectTo": "Yhdistä",
-"Nitrox_AddServerName": "Nimi:",
-"Nitrox_AddServerHost": "Isäntä:",
-"Nitrox_AddServerPort": "Portti:",
-"Nitrox_AddServerAdd": "Lisää",
-"Nitrox_JoinServerPasswordHeader": "Palvelin vaatii salasanan",
-"Nitrox_JoinServerIpAddress": "Palvelimen IP osoite",
-"Nitrox_JoinServerPassword": "Salasana:",
-"Nitrox_JoinServerFailed": "Palvelimelle ei pystynyt liittymään",
-"Nitrox_Cancel": "Hylkää",
-"Nitrox_Join": "Liity",
-"Nitrox_PlayerColor": "Pelaaja väri",
-"Nitrox_EnterName": "Pelaaja nimi",
-"Nitrox_SubmitPassword": "Syötä salasana",
-"Nitrox_MultiplayerLoaded": "Moninpeli ladattu...",
-"Nitrox_StreamerSettings": "Streamer-tila",
-"Nitrox_SilenceChat": "Mykistä chat",
-"Nitrox_HideIp": "Piilota IP-osoitteet",
-// In game menu
-"Nitrox_LoadingScreenWarn": "Odota peliä rikkovia bugeja ja pelin kaatumista",
-"Nitrox_WorldSettling": "Odotetaan mailman asettumista",
-"Nitrox_JoiningSession": "Liitytään moninpeli istuntoon",
-"Nitrox_SyncingWorld": "Synkataan moninpeli maailmaa",
-"Nitrox_LostConnection": "Yhteys menetetty peli palvelimeen",
-"Nitrox_PlayerKicked": "Sinut on potkittu peli palvelimelta",
-// In game logging
-"Nitrox_UnableToConnect": "Etäpalvelimeen ei voitu ottaa yhteyttä osoitteessa:",
-"Nitrox_StartServer": "Käynnistä palvelin ensiksi niin voit liittyä itse hostaamaan maailmaasi",
-"Nitrox_FirewallInterfering": "Näyttää siltä että palomuuri asetuksesi häiritsevät palvelinta",
-"Nitrox_InvalidUserName": "Syötä kelvollinen pelaaja nimi !\n\n Pelaaja nimi ei saa sisältää välilyöntejä tai merkkejä \n Sallitut merkit : A-Z a-z 0-9 _ . -\nLength : [3, 25]",
-"Nitrox_RequestingSessionPolicy": "Istuntokäytäntötietojen pyytäminen...",
-"Nitrox_RejectedSessionPolicy": "Varaus hylätty...",
-"Nitrox_DisconnectedSession": "Yhteys palvelimeen on katkaistu",
-"Nitrox_Waiting": "Odotetaan liittymisjonossa...",
-"Nitrox_WaitingPassword": "Odotetaan palvelimen salasanan syöttöä...",
-"Nitrox_WaitingUserInput": "Odotetaan käyttäjän syöttöä...",
-"Nitrox_LaunchGame": "Käynnistetään peliä...",
-"Nitrox_SilencedChatNotif": "Keskustelu on nyt mykistetty"
+    "Nitrox_Multiplayer": "Moninpeli",
+    "Nitrox_AddServer": "Lisää palvelin",
+    "Nitrox_ConnectTo": "Yhdistä",
+    "Nitrox_AddServerName": "Nimi:",
+    "Nitrox_AddServerHost": "Isäntä:",
+    "Nitrox_AddServerPort": "Portti:",
+    "Nitrox_AddServerAdd": "Lisää",
+    "Nitrox_JoinServerPasswordHeader": "Palvelin vaatii salasanan",
+    "Nitrox_JoinServerIpAddress": "Palvelimen IP osoite",
+    "Nitrox_JoinServerPassword": "Salasana:",
+    "Nitrox_JoinServerFailed": "Palvelimelle ei pystynyt liittymään",
+    "Nitrox_Cancel": "Hylkää",
+    "Nitrox_Join": "Liity",
+    "Nitrox_PlayerColor": "Pelaaja väri",
+    "Nitrox_EnterName": "Pelaaja nimi",
+    "Nitrox_SubmitPassword": "Syötä salasana",
+    "Nitrox_MultiplayerLoaded": "Moninpeli ladattu…",
+    "Nitrox_StreamerSettings": "Streamer-tila",
+    "Nitrox_SilenceChat": "Mykistä chat",
+    "Nitrox_HideIp": "Piilota IP-osoitteet",
+    "Nitrox_LoadingScreenWarn": "Odota peliä rikkovia bugeja ja pelin kaatumista",
+    "Nitrox_WorldSettling": "Odotetaan mailman asettumista",
+    "Nitrox_JoiningSession": "Liitytään moninpeli istuntoon",
+    "Nitrox_SyncingWorld": "Synkataan moninpeli maailmaa",
+    "Nitrox_LostConnection": "Yhteys menetetty peli palvelimeen",
+    "Nitrox_PlayerKicked": "Sinut on potkittu peli palvelimelta",
+    "Nitrox_UnableToConnect": "Etäpalvelimeen ei voitu ottaa yhteyttä osoitteessa:",
+    "Nitrox_StartServer": "Käynnistä palvelin ensiksi niin voit liittyä itse hostaamaan maailmaasi",
+    "Nitrox_FirewallInterfering": "Näyttää siltä että palomuuri asetuksesi häiritsevät palvelinta",
+    "Nitrox_InvalidUserName": "Syötä kelvollinen pelaaja nimi !\n\n Pelaaja nimi ei saa sisältää välilyöntejä tai merkkejä \n Sallitut merkit : A-Z a-z 0-9 _ . -\nLength : [3, 25]",
+    "Nitrox_RequestingSessionPolicy": "Istuntokäytäntötietojen pyytäminen…",
+    "Nitrox_RejectedSessionPolicy": "Varaus hylätty…",
+    "Nitrox_DisconnectedSession": "Yhteys palvelimeen on katkaistu",
+    "Nitrox_Waiting": "Odotetaan liittymisjonossa…",
+    "Nitrox_WaitingPassword": "Odotetaan palvelimen salasanan syöttöä…",
+    "Nitrox_WaitingUserInput": "Odotetaan käyttäjän syöttöä…",
+    "Nitrox_LaunchGame": "Käynnistetään peliä…",
+    "Nitrox_SilencedChatNotif": "Keskustelu on nyt mykistetty"
 }
