+++ Nitrox.Assets.Subnautica/LanguageFiles/German.json
@@ -1,49 +1,47 @@
 {
-"version": 4,
-// Main menu
-"Nitrox_Multiplayer": "Mehrspieler",
-"Nitrox_AddServer": "Server hinzufügen",
-"Nitrox_ConnectTo": "Verbinde zu",
-"Nitrox_AddServerName": "Name:",
-"Nitrox_AddServerHost": "Host:",
-"Nitrox_AddServerPort": "Port:",
-"Nitrox_AddServerAdd": "Hinzufügen",
-"Nitrox_JoinServerPasswordHeader": "Serverpasswort benötigt",
-"Nitrox_JoinServerIpAddress": "Server IP Adresse",
-"Nitrox_JoinServerPassword": "Passwort:",
-"Nitrox_JoinServerFailed": "Die Verbindung zum Server ist gescheitert.",
-"Nitrox_Cancel": "Abbrechen",
-"Nitrox_Join": "Beitreten",
-"Nitrox_PlayerColor": "Spieler Farbe",
-"Nitrox_EnterName": "Geb den Spielername ein",
-"Nitrox_SubmitPassword": "Bestätige Passwort",
-"Nitrox_MultiplayerLoaded": "Mehrspieler Client wurde gestartet...",
-"Nitrox_StreamerSettings": "Streamer Einstellungen",
-"Nitrox_SilenceChat": "Chat stummschalten",
-"Nitrox_HideIp": "IPs verstecken",
-// In game menu
-"Nitrox_LoadingScreenWarn": "Erwarte spielbrechende Fehler",
-"Nitrox_WorldSettling": "Erwarte Fertigstellung der Welt",
-"Nitrox_JoiningSession": "Betrete Mehrspieler Welt",
-"Nitrox_SyncingWorld": "Syncronisiere Mehrspieler Welt",
-"Nitrox_LostConnection": "Die Verbindung zum Server wurde unterbrochen",
-"Nitrox_PlayerKicked": "Du bist vom Server geflogen",
-// In game logging
-"Nitrox_UnableToConnect": "Der Server kann nicht erreicht werden:",
-"Nitrox_StartServer": "Bitte starte zuerst den Server um deine selbst-gehostete Welt zu betreten.",
-"Nitrox_FirewallInterfering": "Es scheint als ob Firewall-Einstellungen die Verbindung blockieren",
-"Nitrox_InvalidUserName": "Bitte gebe einen validen Spielernamen ein!\n\n Er darf kein Leerzeichen oder andere invalide Zeichen beinhalten\n Erlaubte Zeichen : A-Z a-z 0-9 _ . -\nLänge : [3, 25]",
-"Nitrox_RequestingSessionPolicy": "Frage nach Inforamtionen zu Sitzungrichtlinien...",
-"Nitrox_RejectedSessionPolicy": "Anfrage wurde abgelehnt...",
-"Nitrox_DisconnectedSession": "Die Verbindung wurde unterbrochen",
-"Nitrox_Waiting": "Warte in der Beitrittswarteschlange...",
-"Nitrox_WaitingPassword": "Warte auf Eingabe des Server Passworts...",
-"Nitrox_WaitingUserInput": "Warte auf Nutzereingaben...",
-"Nitrox_LaunchGame": "Starte das Spiel...",
-"Nitrox_SilencedChatNotif": "Chat ist stummgeschaltet",
-"Nitrox_CommandNotAvailable": "Dieser Befehl ist in Nitrox nicht verfügbar",
-"Nitrox_DiscordAccept": "Akzeptieren",
-"Nitrox_DiscordDecline": "Ablehnen",
-"Nitrox_DiscordPressToFocus": "Drücke zum interagieren",
-"Nitrox_DiscordRequestText": "Fragt danach deinem Server beizutreten"
+    "Nitrox_Multiplayer": "Mehrspieler",
+    "Nitrox_AddServer": "Server hinzufügen",
+    "Nitrox_ConnectTo": "Verbinde zu",
+    "Nitrox_AddServerName": "Name:",
+    "Nitrox_AddServerHost": "Host:",
+    "Nitrox_AddServerPort": "Port:",
+    "Nitrox_AddServerAdd": "Hinzufügen",
+    "Nitrox_JoinServerPasswordHeader": "Serverpasswort benötigt",
+    "Nitrox_JoinServerIpAddress": "Server IP Adresse",
+    "Nitrox_JoinServerPassword": "Passwort:",
+    "Nitrox_JoinServerFailed": "Die Verbindung zum Server ist gescheitert",
+    "Nitrox_Cancel": "Abbrechen",
+    "Nitrox_Join": "Beitreten",
+    "Nitrox_PlayerColor": "Spieler Farbe",
+    "Nitrox_EnterName": "Geb den Spielername ein",
+    "Nitrox_SubmitPassword": "Bestätige Passwort",
+    "Nitrox_MultiplayerLoaded": "Mehrspieler Client wurde gestartet…",
+    "Nitrox_StreamerSettings": "Streamer Einstellungen",
+    "Nitrox_SilenceChat": "Chat stummschalten",
+    "Nitrox_HideIp": "IPs verstecken",
+    "Nitrox_LoadingScreenWarn": "Erwarte spielbrechende Fehler",
+    "Nitrox_WorldSettling": "Erwarte Fertigstellung der Welt",
+    "Nitrox_JoiningSession": "Betrete Mehrspieler Welt",
+    "Nitrox_SyncingWorld": "Syncronisiere Mehrspieler Welt",
+    "Nitrox_LostConnection": "Die Verbindung zum Server wurde unterbrochen",
+    "Nitrox_PlayerKicked": "Du bist vom Server geflogen",
+    "Nitrox_UnableToConnect": "Der Server kann nicht erreicht werden:",
+    "Nitrox_StartServer": "Bitte starte zuerst den Server um deine selbst-gehostete Welt zu betreten",
+    "Nitrox_FirewallInterfering": "Es scheint als ob Firewall-Einstellungen die Verbindung blockieren",
+    "Nitrox_InvalidUserName": "Bitte gebe einen validen Spielernamen ein!\n\n Er darf kein Leerzeichen oder andere invalide Zeichen beinhalten\n Erlaubte Zeichen : A-Z a-z 0-9 _ . -\nLänge : [3, 25]",
+    "Nitrox_RequestingSessionPolicy": "Frage nach Inforamtionen zu Sitzungrichtlinien…",
+    "Nitrox_RejectedSessionPolicy": "Anfrage wurde abgelehnt…",
+    "Nitrox_DisconnectedSession": "Die Verbindung wurde unterbrochen",
+    "Nitrox_Waiting": "Warte in der Beitrittswarteschlange…",
+    "Nitrox_WaitingPassword": "Warte auf Eingabe des Server Passworts…",
+    "Nitrox_WaitingUserInput": "Warte auf Nutzereingaben…",
+    "Nitrox_LaunchGame": "Starte das Spiel…",
+    "Nitrox_SilencedChatNotif": "Chat ist stummgeschaltet",
+    "Nitrox_CommandNotAvailable": "Dieser Befehl ist in Nitrox nicht verfügbar",
+    "Nitrox_DiscordAccept": "Akzeptieren",
+    "Nitrox_DiscordDecline": "Ablehnen",
+    "Nitrox_DiscordPressToFocus": "Drücke zum interagieren",
+    "Nitrox_DiscordRequestText": "Fragt danach deinem Server beizutreten",
+    "Nitrox_ServerStopped": "Server gestoppt",
+    "Nitrox_LostConnectionWarn": "Nicht länger verbunden mit dem Server"
 }
