+++ Nitrox.Assets.Subnautica/LanguageFiles/Hungarian.json
@@ -1,44 +1,40 @@
 ﻿{
-"version": 3,
-// Main menu
-"Nitrox_Multiplayer": "Többjátékos mód",
-"Nitrox_AddServer": "Szerver hozzáadása",
-"Nitrox_ConnectTo": "Csatlakozás",
-"Nitrox_AddServerName": "Név:",
-"Nitrox_AddServerHost": "Hoszt:",
-"Nitrox_AddServerPort": "Port:",
-"Nitrox_AddServerAdd": "Hozzáad",
-"Nitrox_JoinServerPasswordHeader": "A belépéshez jelszó kell",
-"Nitrox_JoinServerIpAddress": "Szerver IP cím",
-"Nitrox_JoinServerPassword": "Jelszó:",
-"Nitrox_JoinServerFailed": "Nem sikerült csatlakozni a szerverre",
-"Nitrox_Cancel": "Vissza",
-"Nitrox_Join": "Csatlakozás",
-"Nitrox_PlayerColor": "Játékos színe",
-"Nitrox_EnterName": "Játékos neve",
-"Nitrox_SubmitPassword": "Jelszó megadása",
-"Nitrox_MultiplayerLoaded": "Többjátékos Kliens betöltött...",
-"Nitrox_StreamerSettings": "Streamer mód",
-"Nitrox_SilenceChat": "Csevegés elnémítása",
-"Nitrox_HideIp": "IP-címek elrejtése",
-// In game menu
-"Nitrox_LoadingScreenWarn": "Játékélményt romboló hibák várhatóak",
-"Nitrox_WorldSettling": "A világ rendelkezésre vár",
-"Nitrox_JoiningSession": "Betöltés többjátékos munkamenetre",
-"Nitrox_SyncingWorld": "Többjátékos világ szinkronizálása",
-"Nitrox_LostConnection": "Elveszett a kapcsolat a többjátékos szerverrel",
-"Nitrox_PlayerKicked": "Ki lettél rúgva a szerverről",
-// In game logging
-"Nitrox_UnableToConnect": "Nem sikerült csatlakozni a megadott távoli szerverre:",
-"Nitrox_StartServer": "Először indítsd el a szervered, hogy csatlakozni tudj a saját szerveredre",
-"Nitrox_FirewallInterfering": "Kérlek állítsd át a Tűzfal beállításaidat, hogy működjön a játék",
-"Nitrox_InvalidUserName": "Kérlek, érvényes nevet adj meg !\n\n Nem tartalmazhat szóközt, vagy kétes karaktereket\n Megengetett karakterek : A-Z a-z 0-9 _ . -\nHosszúság : [3, 25]",
-"Nitrox_RequestingSessionPolicy": "Munkamenet irányelvek lekérése...",
-"Nitrox_RejectedSessionPolicy": "A foglalás elutasítva...",
-"Nitrox_DisconnectedSession": "Kilépve a munkameneből",
-"Nitrox_Waiting": "Várakozás a várólistán",
-"Nitrox_WaitingPassword": "Várakozás a szerver jelszó megadására...",
-"Nitrox_WaitingUserInput": "Várakozás a felhasználóra...",
-"Nitrox_LaunchGame": "Játék indítása...",
-"Nitrox_SilencedChatNotif": "A csevegés mostantól néma"
+    "Nitrox_Multiplayer": "Többjátékos mód",
+    "Nitrox_AddServer": "Szerver hozzáadása",
+    "Nitrox_ConnectTo": "Csatlakozás",
+    "Nitrox_AddServerName": "Név:",
+    "Nitrox_AddServerHost": "Hoszt:",
+    "Nitrox_AddServerPort": "Port:",
+    "Nitrox_AddServerAdd": "Hozzáad",
+    "Nitrox_JoinServerPasswordHeader": "A belépéshez jelszó kell",
+    "Nitrox_JoinServerIpAddress": "Szerver IP cím",
+    "Nitrox_JoinServerPassword": "Jelszó:",
+    "Nitrox_JoinServerFailed": "Nem sikerült csatlakozni a szerverre",
+    "Nitrox_Cancel": "Vissza",
+    "Nitrox_Join": "Csatlakozás",
+    "Nitrox_PlayerColor": "Játékos színe",
+    "Nitrox_EnterName": "Játékos neve",
+    "Nitrox_SubmitPassword": "Jelszó megadása",
+    "Nitrox_MultiplayerLoaded": "Többjátékos Kliens betöltött…",
+    "Nitrox_StreamerSettings": "Streamer mód",
+    "Nitrox_SilenceChat": "Csevegés elnémítása",
+    "Nitrox_HideIp": "IP-címek elrejtése",
+    "Nitrox_LoadingScreenWarn": "Játékélményt romboló hibák várhatóak",
+    "Nitrox_WorldSettling": "A világ rendelkezésre vár",
+    "Nitrox_JoiningSession": "Betöltés többjátékos munkamenetre",
+    "Nitrox_SyncingWorld": "Többjátékos világ szinkronizálása",
+    "Nitrox_LostConnection": "Elveszett a kapcsolat a többjátékos szerverrel",
+    "Nitrox_PlayerKicked": "Ki lettél rúgva a szerverről",
+    "Nitrox_UnableToConnect": "Nem sikerült csatlakozni a megadott távoli szerverre:",
+    "Nitrox_StartServer": "Először indítsd el a szervered, hogy csatlakozni tudj a saját szerveredre",
+    "Nitrox_FirewallInterfering": "Kérlek állítsd át a Tűzfal beállításaidat, hogy működjön a játék",
+    "Nitrox_InvalidUserName": "Kérlek, érvényes nevet adj meg !\n\n Nem tartalmazhat szóközt, vagy kétes karaktereket\n Megengetett karakterek : A-Z a-z 0-9 _ . -\nHosszúság : [3, 25]",
+    "Nitrox_RequestingSessionPolicy": "Munkamenet irányelvek lekérése…",
+    "Nitrox_RejectedSessionPolicy": "A foglalás elutasítva…",
+    "Nitrox_DisconnectedSession": "Kilépve a munkameneből",
+    "Nitrox_Waiting": "Várakozás a várólistán…",
+    "Nitrox_WaitingPassword": "Várakozás a szerver jelszó megadására…",
+    "Nitrox_WaitingUserInput": "Várakozás a felhasználóra…",
+    "Nitrox_LaunchGame": "Játék indítása…",
+    "Nitrox_SilencedChatNotif": "A csevegés mostantól néma"
 }
