+++ Nitrox.Assets.Subnautica/LanguageFiles/Italian.json
@@ -1,45 +1,41 @@
 {
-"version": 4,
-// Main menu
-"Nitrox_Multiplayer": "Multiplayer",
-"Nitrox_AddServer": "Aggiungi un server",
-"Nitrox_ConnectTo": "Connettiti a",
-"Nitrox_AddServerName": "Nome:",
-"Nitrox_AddServerHost": "Indirizzo IP:",
-"Nitrox_AddServerPort": "Porta:",
-"Nitrox_AddServerAdd": "Aggiungi",
-"Nitrox_JoinServerPasswordHeader": "Richiesta password",
-"Nitrox_JoinServerIpAddress": "Indirizzo IP",
-"Nitrox_JoinServerPassword": "Password:",
-"Nitrox_JoinServerFailed": "Impossibile entrare nel server",
-"Nitrox_Cancel": "Annulla",
-"Nitrox_Join": "Entra",
-"Nitrox_PlayerColor": "Colore giocatore",
-"Nitrox_EnterName": "Nome giocatore",
-"Nitrox_SubmitPassword": "Conferma",
-"Nitrox_MultiplayerLoaded": "Client Multiplayer caricato...",
-"Nitrox_StreamerSettings": "Modalità Streamer",
-"Nitrox_SilenceChat": "Disattiva la chat",
-"Nitrox_HideIp": "Nascondere gli indirizzi IP",
-// In game menu
-"Nitrox_LoadingScreenWarn": "Aspettati problemi che potrebbero compromettere l'esperienza di gioco",
-"Nitrox_WorldSettling": "Caricamento mondo di gioco",
-"Nitrox_JoiningSession": "Sto entrando nella sessione Multiplayer",
-"Nitrox_SyncingWorld": "Sincronizzazione mondo di gioco",
-"Nitrox_LostConnection": "Connessione al server persa",
-"Nitrox_PlayerKicked": "Sei stato espulso dal server",
-// In game logging
-"Nitrox_UnableToConnect": "Impossibile contattare il server ",
-"Nitrox_StartServer": "Per entrare nel tuo server locale devi prima avviarlo da Nitrox Launcher",
-"Nitrox_FirewallInterfering": "Sembrerebbe che il firewall stia interferendo",
-"Nitrox_InvalidUserName": "Per favore inserisce un nome valido!\n\n Non può contenere spazi o caratteri ambigui\n Caratteri consentiti : A-Z a-z 0-9 _ . -\nLunghezza : [3, 25]",
-"Nitrox_RequestingSessionPolicy": "Sto richiedendo le impostazioni della sessione...",
-"Nitrox_RejectedSessionPolicy": "Richiesta rifiutata...",
-"Nitrox_DisconnectedSession": "Disconnesso dal server",
-"Nitrox_Waiting": "In attesa in coda di iscrizione...",
-"Nitrox_WaitingPassword": "In attesa della password per il server...",
-"Nitrox_WaitingUserInput": "In attesa dell'inserimento dal parte dell'utente...",
-"Nitrox_LaunchGame": "Caricamento...",
-"Nitrox_SilencedChatNotif": "La chat è ora silenziata",
-"Nitrox_CommandNotAvailable": "Questo comando non è disponibile con Nitrox"
+    "Nitrox_Multiplayer": "Multiplayer",
+    "Nitrox_AddServer": "Aggiungi un server",
+    "Nitrox_ConnectTo": "Connettiti a",
+    "Nitrox_AddServerName": "Nome:",
+    "Nitrox_AddServerHost": "Indirizzo IP:",
+    "Nitrox_AddServerPort": "Porta:",
+    "Nitrox_AddServerAdd": "Aggiungi",
+    "Nitrox_JoinServerPasswordHeader": "Richiesta password",
+    "Nitrox_JoinServerIpAddress": "Indirizzo IP",
+    "Nitrox_JoinServerPassword": "Password:",
+    "Nitrox_JoinServerFailed": "Impossibile entrare nel server",
+    "Nitrox_Cancel": "Annulla",
+    "Nitrox_Join": "Entra",
+    "Nitrox_PlayerColor": "Colore giocatore",
+    "Nitrox_EnterName": "Nome giocatore",
+    "Nitrox_SubmitPassword": "Conferma",
+    "Nitrox_MultiplayerLoaded": "Client Multiplayer caricato…",
+    "Nitrox_StreamerSettings": "Modalità Streamer",
+    "Nitrox_SilenceChat": "Disattiva la chat",
+    "Nitrox_HideIp": "Nascondere gli indirizzi IP",
+    "Nitrox_LoadingScreenWarn": "Aspettati problemi che potrebbero compromettere l'esperienza di gioco",
+    "Nitrox_WorldSettling": "Caricamento mondo di gioco",
+    "Nitrox_JoiningSession": "Sto entrando nella sessione Multiplayer",
+    "Nitrox_SyncingWorld": "Sincronizzazione mondo di gioco",
+    "Nitrox_LostConnection": "Connessione al server persa",
+    "Nitrox_PlayerKicked": "Sei stato espulso dal server",
+    "Nitrox_UnableToConnect": "Impossibile contattare il server",
+    "Nitrox_StartServer": "Per entrare nel tuo server locale devi prima avviarlo da Nitrox Launcher",
+    "Nitrox_FirewallInterfering": "Sembrerebbe che il firewall stia interferendo",
+    "Nitrox_InvalidUserName": "Per favore inserisce un nome valido!\n\n Non può contenere spazi o caratteri ambigui\n Caratteri consentiti : A-Z a-z 0-9 _ . -\nLunghezza : [3, 25]",
+    "Nitrox_RequestingSessionPolicy": "Sto richiedendo le impostazioni della sessione…",
+    "Nitrox_RejectedSessionPolicy": "Richiesta rifiutata…",
+    "Nitrox_DisconnectedSession": "Disconnesso dal server",
+    "Nitrox_Waiting": "In attesa in coda di iscrizione…",
+    "Nitrox_WaitingPassword": "In attesa della password per il server…",
+    "Nitrox_WaitingUserInput": "In attesa dell'inserimento dal parte dell'utente…",
+    "Nitrox_LaunchGame": "Caricamento…",
+    "Nitrox_SilencedChatNotif": "La chat è ora silenziata",
+    "Nitrox_CommandNotAvailable": "Questo comando non è disponibile con Nitrox"
 }
