+++ Nitrox.Assets.Subnautica/LanguageFiles/Polish.json
@@ -1,44 +1,40 @@
 {
-"version": 3,
-// Main menu
-"Nitrox_Multiplayer": "Gra wieloosobowa",
-"Nitrox_AddServer": "Dodaj serwer",
-"Nitrox_ConnectTo": "Połącz z",
-"Nitrox_AddServerName": "Nazwa:",
-"Nitrox_AddServerHost": "Adres:",
-"Nitrox_AddServerPort": "Port:",
-"Nitrox_AddServerAdd": "Dodaj",
-"Nitrox_JoinServerPasswordHeader": "Serwer wymaga hasła",
-"Nitrox_JoinServerIpAddress": "IP serwera",
-"Nitrox_JoinServerPassword": "Hasło:",
-"Nitrox_JoinServerFailed": "Dołączenie na serwer się nie powiodło",
-"Nitrox_Cancel": "Anuluj",
-"Nitrox_Join": "Dołącz",
-"Nitrox_PlayerColor": "Kolor gracza",
-"Nitrox_EnterName": "Nazwa gracza",
-"Nitrox_SubmitPassword": "Potwierdź",
-"Nitrox_MultiplayerLoaded": "Klient wieloosobowy wczytany...",
-"Nitrox_StreamerSettings": "Tryb Streamera",
-"Nitrox_SilenceChat": "Wycisz czat",
-"Nitrox_HideIp": "Ukryj adresy IP",
-// In game menu
-"Nitrox_LoadingScreenWarn": "Spodziewaj się błędów utrudniających rozgrywkę",
-"Nitrox_WorldSettling": "Oczekiwanie na osadzenie świata",
-"Nitrox_JoiningSession": "Dołączanie do sesji wieloosobowej",
-"Nitrox_SyncingWorld": "Synchronizacja świata",
-"Nitrox_LostConnection": "Utracono połączenie z serwerem gry",
-"Nitrox_PlayerKicked": "Wyrzucono Cię z serwera",
-// In game logging
-"Nitrox_UnableToConnect": "Nie można połączyć się ze zdalnym serwerem:",
-"Nitrox_StartServer": "Najpierw uruchom serwer aby dołączyć do swojego świata",
-"Nitrox_FirewallInterfering": "Wygląda na to, że zapora sieciowa blokuje połączenie",
-"Nitrox_InvalidUserName": "Wpisz poprawną nazwę gracza! \n\n Nie może zawierać żadnych spacji ani podobnych znaków\nDozwolone znaki: A-Z a-z 0-9 _ . -\nDługość: [3, 25]",
-"Nitrox_RequestingSessionPolicy": "Żądanie informacji o warunkach sesji...",
-"Nitrox_RejectedSessionPolicy": "Rezerwacja odrzucona...",
-"Nitrox_DisconnectedSession": "Odłączono od serwera",
-"Nitrox_Waiting": "Oczekiwanie w kolejce do dołączenia",
-"Nitrox_WaitingPassword": "Oczekiwanie na wprowadzenie hasła do serwera...",
-"Nitrox_WaitingUserInput": "Oczekiwanie na dane od użytkownika...",
-"Nitrox_LaunchGame": "Uruchamianie gry...",
-"Nitrox_SilencedChatNotif": "Czat jest teraz wyciszony"
+    "Nitrox_Multiplayer": "Gra wieloosobowa",
+    "Nitrox_AddServer": "Dodaj serwer",
+    "Nitrox_ConnectTo": "Połącz z",
+    "Nitrox_AddServerName": "Nazwa:",
+    "Nitrox_AddServerHost": "Adres:",
+    "Nitrox_AddServerPort": "Port:",
+    "Nitrox_AddServerAdd": "Dodaj",
+    "Nitrox_JoinServerPasswordHeader": "Serwer wymaga hasła",
+    "Nitrox_JoinServerIpAddress": "IP serwera",
+    "Nitrox_JoinServerPassword": "Hasło:",
+    "Nitrox_JoinServerFailed": "Dołączenie na serwer się nie powiodło",
+    "Nitrox_Cancel": "Anuluj",
+    "Nitrox_Join": "Dołącz",
+    "Nitrox_PlayerColor": "Kolor gracza",
+    "Nitrox_EnterName": "Nazwa gracza",
+    "Nitrox_SubmitPassword": "Potwierdź",
+    "Nitrox_MultiplayerLoaded": "Klient wieloosobowy wczytany…",
+    "Nitrox_StreamerSettings": "Tryb Streamera",
+    "Nitrox_SilenceChat": "Wycisz czat",
+    "Nitrox_HideIp": "Ukryj adresy IP",
+    "Nitrox_LoadingScreenWarn": "Spodziewaj się błędów utrudniających rozgrywkę",
+    "Nitrox_WorldSettling": "Oczekiwanie na osadzenie świata",
+    "Nitrox_JoiningSession": "Dołączanie do sesji wieloosobowej",
+    "Nitrox_SyncingWorld": "Synchronizacja świata",
+    "Nitrox_LostConnection": "Utracono połączenie z serwerem gry",
+    "Nitrox_PlayerKicked": "Wyrzucono Cię z serwera",
+    "Nitrox_UnableToConnect": "Nie można połączyć się ze zdalnym serwerem:",
+    "Nitrox_StartServer": "Najpierw uruchom serwer aby dołączyć do swojego świata",
+    "Nitrox_FirewallInterfering": "Wygląda na to, że zapora sieciowa blokuje połączenie",
+    "Nitrox_InvalidUserName": "Wpisz poprawną nazwę gracza! \n\n Nie może zawierać żadnych spacji ani podobnych znaków\nDozwolone znaki: A-Z a-z 0-9 _ . -\nDługość: [3, 25]",
+    "Nitrox_RequestingSessionPolicy": "Żądanie informacji o warunkach sesji…",
+    "Nitrox_RejectedSessionPolicy": "Rezerwacja odrzucona…",
+    "Nitrox_DisconnectedSession": "Odłączono od serwera",
+    "Nitrox_Waiting": "Oczekiwanie w kolejce do dołączenia…",
+    "Nitrox_WaitingPassword": "Oczekiwanie na wprowadzenie hasła do serwera…",
+    "Nitrox_WaitingUserInput": "Oczekiwanie na dane od użytkownika…",
+    "Nitrox_LaunchGame": "Uruchamianie gry…",
+    "Nitrox_SilencedChatNotif": "Czat jest teraz wyciszony"
 }
