+++ Nitrox.Assets.Subnautica/LanguageFiles/Portuguese.json
@@ -1,45 +1,41 @@
 {
-"version": 4,
-// Main menu
-"Nitrox_Multiplayer": "Multijugador",
-"Nitrox_AddServer": "Adicionar Servidor",
-"Nitrox_ConnectTo": "Conectar",
-"Nitrox_AddServerName": "Nome:",
-"Nitrox_AddServerHost": "Host:",
-"Nitrox_AddServerPort": "Porta:",
-"Nitrox_AddServerAdd": "Adicionar",
-"Nitrox_JoinServerPasswordHeader": "Senha do servidor necessaria",
-"Nitrox_JoinServerIpAddress": "IP do servidor",
-"Nitrox_JoinServerPassword": "Senha:",
-"Nitrox_JoinServerFailed": "Nao foi possivel entrar no servidor",
-"Nitrox_Cancel": "Cancelar",
-"Nitrox_Join": "Junte-se",
-"Nitrox_PlayerColor": "Cor do jogador",
-"Nitrox_EnterName": "Nome do jogador",
-"Nitrox_SubmitPassword": "Enviar",
-"Nitrox_MultiplayerLoaded": "Cliente multijogador carregado...",
-"Nitrox_StreamerSettings": "Modo streamer",
-"Nitrox_SilenceChat": "Silenciar a conversa",
-"Nitrox_HideIp": "Ocultar IPs",
-// In game menu
-"Nitrox_LoadingScreenWarn": "Espere bugs que quebram o jogo",
-"Nitrox_WorldSettling": "Aguardando o estabelecimento do mundo",
-"Nitrox_JoiningSession": "Entrando na Sessao Multijogador",
-"Nitrox_SyncingWorld": "Sincronizando Mundo Multijogador",
-"Nitrox_LostConnection": "Conexao perdida com o servidor do jogo",
-"Nitrox_PlayerKicked": "Voce foi expulso do servidor",
-// In game logging
-"Nitrox_UnableToConnect": "Nao e possivel entrar em contato com o servidor remoto em:",
-"Nitrox_StartServer": "Inicie o seu servidor primeiro para se juntar ao seu mundo auto-hospedado",
-"Nitrox_FirewallInterfering": "Parece que as configuracoes do seu firewall estao interferindo",
-"Nitrox_InvalidUserName": "Por favor, digite um nome de jogador valido!\n\n Nao deve conter nenhum espaço ou caracteres duvidosos\n Caracteres permitidos : A-Z a-z 0-9 _ . -\nComprimento : [3, 25]",
-"Nitrox_RequestingSessionPolicy": "Solicitando informacoes de política de sessao...",
-"Nitrox_RejectedSessionPolicy": "Reserva rejeitada...",
-"Nitrox_DisconnectedSession": "Desconectado do servidor",
-"Nitrox_Waiting": "Esperando na fila de entrada...",
-"Nitrox_WaitingPassword": "Aguardando entrada de senha do servidor...",
-"Nitrox_WaitingUserInput": "Aguardando entrada do usuaio...",
-"Nitrox_LaunchGame": "Iniciando o jogo...",
-"Nitrox_SilencedChatNotif": "A conversa é silenciada",
-"Nitrox_CommandNotAvailable": "Este comando não está disponível com Nitrox"
+    "Nitrox_Multiplayer": "Multijugador",
+    "Nitrox_AddServer": "Adicionar Servidor",
+    "Nitrox_ConnectTo": "Conectar",
+    "Nitrox_AddServerName": "Nome:",
+    "Nitrox_AddServerHost": "Host:",
+    "Nitrox_AddServerPort": "Porta:",
+    "Nitrox_AddServerAdd": "Adicionar",
+    "Nitrox_JoinServerPasswordHeader": "Senha do servidor necessaria",
+    "Nitrox_JoinServerIpAddress": "IP do servidor",
+    "Nitrox_JoinServerPassword": "Senha:",
+    "Nitrox_JoinServerFailed": "Nao foi possivel entrar no servidor",
+    "Nitrox_Cancel": "Cancelar",
+    "Nitrox_Join": "Junte-se",
+    "Nitrox_PlayerColor": "Cor do jogador",
+    "Nitrox_EnterName": "Nome do jogador",
+    "Nitrox_SubmitPassword": "Enviar",
+    "Nitrox_MultiplayerLoaded": "Cliente multijogador carregado…",
+    "Nitrox_StreamerSettings": "Modo streamer",
+    "Nitrox_SilenceChat": "Silenciar a conversa",
+    "Nitrox_HideIp": "Ocultar IPs",
+    "Nitrox_LoadingScreenWarn": "Espere bugs que quebram o jogo",
+    "Nitrox_WorldSettling": "Aguardando o estabelecimento do mundo",
+    "Nitrox_JoiningSession": "Entrando na Sessao Multijogador",
+    "Nitrox_SyncingWorld": "Sincronizando Mundo Multijogador",
+    "Nitrox_LostConnection": "Conexao perdida com o servidor do jogo",
+    "Nitrox_PlayerKicked": "Voce foi expulso do servidor",
+    "Nitrox_UnableToConnect": "Nao e possivel entrar em contato com o servidor remoto em:",
+    "Nitrox_StartServer": "Inicie o seu servidor primeiro para se juntar ao seu mundo auto-hospedado",
+    "Nitrox_FirewallInterfering": "Parece que as configuracoes do seu firewall estao interferindo",
+    "Nitrox_InvalidUserName": "Por favor, digite um nome de jogador valido!\n\n Nao deve conter nenhum espaço ou caracteres duvidosos\n Caracteres permitidos : A-Z a-z 0-9 _ . -\nComprimento : [3, 25]",
+    "Nitrox_RequestingSessionPolicy": "Solicitando informacoes de política de sessao…",
+    "Nitrox_RejectedSessionPolicy": "Reserva rejeitada…",
+    "Nitrox_DisconnectedSession": "Desconectado do servidor",
+    "Nitrox_Waiting": "Esperando na fila de entrada…",
+    "Nitrox_WaitingPassword": "Aguardando entrada de senha do servidor…",
+    "Nitrox_WaitingUserInput": "Aguardando entrada do usuaio…",
+    "Nitrox_LaunchGame": "Iniciando o jogo…",
+    "Nitrox_SilencedChatNotif": "A conversa é silenciada",
+    "Nitrox_CommandNotAvailable": "Este comando não está disponível com Nitrox"
 }
