+++ Nitrox.Assets.Subnautica/LanguageFiles/Romanian.json
@@ -1,44 +1,40 @@
 {
-"version": 3,
-// Main menu
-"Nitrox_Multiplayer": "Multiplayer",
-"Nitrox_AddServer": "Adaugă server",
-"Nitrox_ConnectTo": "Conectează la",
-"Nitrox_AddServerName": "Nume:",
-"Nitrox_AddServerHost": "Gazdă:",
-"Nitrox_AddServerPort": "Port:",
-"Nitrox_AddServerAdd": "Adăugă",
-"Nitrox_JoinServerPasswordHeader": "Parola serverului este necesară",
-"Nitrox_JoinServerIpAddress": "Adresa IP a serverului",
-"Nitrox_JoinServerPassword": "Parola:",
-"Nitrox_JoinServerFailed": "Nu te poți alătura serverului",
-"Nitrox_Cancel": "Anulare",
-"Nitrox_Join": "Alătură-te",
-"Nitrox_PlayerColor": "Culoarea jucătorului",
-"Nitrox_EnterName": "Introdu numele de jucător",
-"Nitrox_SubmitPassword": "Trimite parola",
-"Nitrox_MultiplayerLoaded": "Clientul Multiplayer încărcat...",
-"Nitrox_StreamerSettings": "Mod Streamer",
-"Nitrox_SilenceChat": "Ascundeți chat-ul",
-"Nitrox_HideIp": "Ascundeți adresele IP",
-// In game menu
-"Nitrox_LoadingScreenWarn": "Așteptă-te la bug-uri care pot strica jocul",
-"Nitrox_WorldSettling": "În așteptarea stabilirii lumii",
-"Nitrox_JoiningSession": "Alăturarea la sesiune Multiplayer",
-"Nitrox_SyncingWorld": "Sincronizarea Lumii Multiplayer",
-"Nitrox_LostConnection": "Conexiune pierdută la serverul jocului",
-"Nitrox_PlayerKicked": "Ai fost dat afară de pe server",
-// In game logging
-"Nitrox_UnableToConnect": "Nu se poate contacta serverul de la distanță la:",
-"Nitrox_StartServer": "Pornește mai întâi serverul pentru a te alătura lumii găzduite de tine",
-"Nitrox_FirewallInterfering": "Se pare că setările firewall-ului tău interferează",
-"Nitrox_InvalidUserName": "Te rugăm să introduci un nume de jucător valid!\n\n Trebuie să nu conțină spațiu sau caractere ambigue\n Caracterele permise : A-Z a-z 0-9 _ . -\nLungime : [3, 25]",
-"Nitrox_RequestingSessionPolicy": "Se solicită informații despre politica sesiunii...",
-"Nitrox_RejectedSessionPolicy": "Rezervare respinsă...",
-"Nitrox_DisconnectedSession": "Deconectat de la server",
-"Nitrox_Waiting": "Așteptând în coadă de așteptare pentru a se alătura serverului",
-"Nitrox_WaitingPassword": "Se așteaptă introducerea parolei serverului...",
-"Nitrox_WaitingUserInput": "Se așteaptă datele introduse de utilizator...",
-"Nitrox_LaunchGame": "Se lansează jocul...",
-"Nitrox_SilencedChatNotif": "Chat-ul este acum dezactivat"
+    "Nitrox_Multiplayer": "Multiplayer",
+    "Nitrox_AddServer": "Adaugă server",
+    "Nitrox_ConnectTo": "Conectează la",
+    "Nitrox_AddServerName": "Nume:",
+    "Nitrox_AddServerHost": "Gazdă:",
+    "Nitrox_AddServerPort": "Port:",
+    "Nitrox_AddServerAdd": "Adăugă",
+    "Nitrox_JoinServerPasswordHeader": "Parola serverului este necesară",
+    "Nitrox_JoinServerIpAddress": "Adresa IP a serverului",
+    "Nitrox_JoinServerPassword": "Parola:",
+    "Nitrox_JoinServerFailed": "Nu te poți alătura serverului",
+    "Nitrox_Cancel": "Anulare",
+    "Nitrox_Join": "Alătură-te",
+    "Nitrox_PlayerColor": "Culoarea jucătorului",
+    "Nitrox_EnterName": "Introdu numele de jucător",
+    "Nitrox_SubmitPassword": "Trimite parola",
+    "Nitrox_MultiplayerLoaded": "Clientul Multiplayer încărcat…",
+    "Nitrox_StreamerSettings": "Mod Streamer",
+    "Nitrox_SilenceChat": "Ascundeți chat-ul",
+    "Nitrox_HideIp": "Ascundeți adresele IP",
+    "Nitrox_LoadingScreenWarn": "Așteptă-te la bug-uri care pot strica jocul",
+    "Nitrox_WorldSettling": "În așteptarea stabilirii lumii",
+    "Nitrox_JoiningSession": "Alăturarea la sesiune Multiplayer",
+    "Nitrox_SyncingWorld": "Sincronizarea Lumii Multiplayer",
+    "Nitrox_LostConnection": "Conexiune pierdută la serverul jocului",
+    "Nitrox_PlayerKicked": "Ai fost dat afară de pe server",
+    "Nitrox_UnableToConnect": "Nu se poate contacta serverul de la distanță la:",
+    "Nitrox_StartServer": "Pornește mai întâi serverul pentru a te alătura lumii găzduite de tine",
+    "Nitrox_FirewallInterfering": "Se pare că setările firewall-ului tău interferează",
+    "Nitrox_InvalidUserName": "Te rugăm să introduci un nume de jucător valid!\n\n Trebuie să nu conțină spațiu sau caractere ambigue\n Caracterele permise : A-Z a-z 0-9 _ . -\nLungime : [3, 25]",
+    "Nitrox_RequestingSessionPolicy": "Se solicită informații despre politica sesiunii…",
+    "Nitrox_RejectedSessionPolicy": "Rezervare respinsă…",
+    "Nitrox_DisconnectedSession": "Deconectat de la server",
+    "Nitrox_Waiting": "Așteptând în coadă de așteptare pentru a se alătura serverului…",
+    "Nitrox_WaitingPassword": "Se așteaptă introducerea parolei serverului…",
+    "Nitrox_WaitingUserInput": "Se așteaptă datele introduse de utilizator…",
+    "Nitrox_LaunchGame": "Se lansează jocul…",
+    "Nitrox_SilencedChatNotif": "Chat-ul este acum dezactivat"
 }
