+++ Nitrox.Assets.Subnautica/LanguageFiles/Russian.json
@@ -1,44 +1,40 @@
 {
-"version": 3,
-// Main menu
-"Nitrox_Multiplayer": "Сетевая игра",
-"Nitrox_AddServer": "Добавить сервер",
-"Nitrox_ConnectTo": "Подключиться",
-"Nitrox_AddServerName": "Имя сервера:",
-"Nitrox_AddServerHost": "Адрес:",
-"Nitrox_AddServerPort": "Порт:",
-"Nitrox_AddServerAdd": "Добавить",
-"Nitrox_JoinServerPasswordHeader": "Необходим пароль",
-"Nitrox_JoinServerIpAddress": "IP сервера",
-"Nitrox_JoinServerPassword": "Пароль:",
-"Nitrox_JoinServerFailed": "Не удаётся подключится к серверу",
-"Nitrox_Cancel": "Отмена",
-"Nitrox_Join": "Войти",
-"Nitrox_PlayerColor": "Цвет игрока",
-"Nitrox_EnterName": "Имя игрока",
-"Nitrox_SubmitPassword": "Принять",
-"Nitrox_MultiplayerLoaded": "Загрузка сетевой игры...",
-"Nitrox_StreamerSettings": "Режим стримера",
-"Nitrox_SilenceChat": "Отключить звук в чате",
-"Nitrox_HideIp": "Скрыть IP-адреса",
-// In game menu
-"Nitrox_LoadingScreenWarn": "Осторожно, иногда могут возникать ошибки",
-"Nitrox_WorldSettling": "Ожидание настроек сервера",
-"Nitrox_JoiningSession": "Вход в сетевую игру",
-"Nitrox_SyncingWorld": "Синхронизация мира",
-"Nitrox_LostConnection": "Соединение с сервером потеряно",
-"Nitrox_PlayerKicked": "Вы были исключены с сервера",
-// In game logging
-"Nitrox_UnableToConnect": "Невозможно связаться с сервером по адресу:",
-"Nitrox_StartServer": "Сначала запустите свой сервер, чтобы присоединиться к вашему автономному миру",
-"Nitrox_FirewallInterfering": "Некорректные настройки брандмауэра",
-"Nitrox_InvalidUserName": "Пожалуйста введите корректное имя!\n\n Доступные символы: A-Z a-z 0-9 _ . -\nДлинна: [3, 25]",
-"Nitrox_RequestingSessionPolicy": "Запрос информации о политике сеанса...",
-"Nitrox_RejectedSessionPolicy": "Бронирование отклонено...",
-"Nitrox_DisconnectedSession": "Отключение от сервера",
-"Nitrox_Waiting": "Ожидание в очереди на подключение",
-"Nitrox_WaitingPassword": "Ожидание ввода пароля сервера...",
-"Nitrox_WaitingUserInput": "Ожидание ввода данных пользователем...",
-"Nitrox_LaunchGame": "Загрузка игры...",
-"Nitrox_SilencedChatNotif": "Чат теперь отключен"
+    "Nitrox_Multiplayer": "Сетевая игра",
+    "Nitrox_AddServer": "Добавить сервер",
+    "Nitrox_ConnectTo": "Подключиться",
+    "Nitrox_AddServerName": "Имя сервера:",
+    "Nitrox_AddServerHost": "Адрес:",
+    "Nitrox_AddServerPort": "Порт:",
+    "Nitrox_AddServerAdd": "Добавить",
+    "Nitrox_JoinServerPasswordHeader": "Необходим пароль",
+    "Nitrox_JoinServerIpAddress": "IP сервера",
+    "Nitrox_JoinServerPassword": "Пароль:",
+    "Nitrox_JoinServerFailed": "Не удаётся подключится к серверу",
+    "Nitrox_Cancel": "Отмена",
+    "Nitrox_Join": "Войти",
+    "Nitrox_PlayerColor": "Цвет игрока",
+    "Nitrox_EnterName": "Имя игрока",
+    "Nitrox_SubmitPassword": "Принять",
+    "Nitrox_MultiplayerLoaded": "Загрузка сетевой игры…",
+    "Nitrox_StreamerSettings": "Режим стримера",
+    "Nitrox_SilenceChat": "Отключить звук в чате",
+    "Nitrox_HideIp": "Скрыть IP-адреса",
+    "Nitrox_LoadingScreenWarn": "Осторожно, иногда могут возникать ошибки",
+    "Nitrox_WorldSettling": "Ожидание настроек сервера",
+    "Nitrox_JoiningSession": "Вход в сетевую игру",
+    "Nitrox_SyncingWorld": "Синхронизация мира",
+    "Nitrox_LostConnection": "Соединение с сервером потеряно",
+    "Nitrox_PlayerKicked": "Вы были исключены с сервера",
+    "Nitrox_UnableToConnect": "Невозможно связаться с сервером по адресу:",
+    "Nitrox_StartServer": "Сначала запустите свой сервер, чтобы присоединиться к вашему автономному миру",
+    "Nitrox_FirewallInterfering": "Некорректные настройки брандмауэра",
+    "Nitrox_InvalidUserName": "Пожалуйста введите корректное имя!\n\n Доступные символы: A-Z a-z 0-9 _ . -\nДлинна: [3, 25]",
+    "Nitrox_RequestingSessionPolicy": "Запрос информации о политике сеанса…",
+    "Nitrox_RejectedSessionPolicy": "Бронирование отклонено…",
+    "Nitrox_DisconnectedSession": "Отключение от сервера",
+    "Nitrox_Waiting": "Ожидание в очереди на подключение…",
+    "Nitrox_WaitingPassword": "Ожидание ввода пароля сервера…",
+    "Nitrox_WaitingUserInput": "Ожидание ввода данных пользователем…",
+    "Nitrox_LaunchGame": "Загрузка игры…",
+    "Nitrox_SilencedChatNotif": "Чат теперь отключен"
 }
