+++ Nitrox.Assets.Subnautica/LanguageFiles/Slovak.json
@@ -1,44 +1,40 @@
 {
-"version": 3,
-// Main menu
-"Nitrox_Multiplayer": "Hrať spoločne",
-"Nitrox_AddServer": "Pridať server",
-"Nitrox_ConnectTo": "Pripojiť sa na",
-"Nitrox_AddServerName": "Názov:",
-"Nitrox_AddServerHost": "Adresa:",
-"Nitrox_AddServerPort": "Port:",
-"Nitrox_AddServerAdd": "Pridať",
-"Nitrox_JoinServerPasswordHeader": "Server požaduje heslo",
-"Nitrox_JoinServerIpAddress": "IP adresa servera",
-"Nitrox_JoinServerPassword": "Heslo:",
-"Nitrox_JoinServerFailed": "Nedá sa pripojiť na server",
-"Nitrox_Cancel": "Zrušiť",
-"Nitrox_Join": "Pripojiť sa",
-"Nitrox_PlayerColor": "Farba hráča",
-"Nitrox_EnterName": "Meno hráča",
-"Nitrox_SubmitPassword": "Použiť",
-"Nitrox_MultiplayerLoaded": "Multiplayer klient načítaný...",
-"Nitrox_StreamerSettings": "Režim streamer",
-"Nitrox_SilenceChat": "Vypnutie chatu",
-"Nitrox_HideIp": "Skryť IP adresy",
-// In game menu
-"Nitrox_LoadingScreenWarn": "Očakávajte chyby, ktoré kazia hru",
-"Nitrox_WorldSettling": "Čakám na nastavenia sveta",
-"Nitrox_JoiningSession": "Pripájam sa do hry",
-"Nitrox_SyncingWorld": "Synchronizujem svet",
-"Nitrox_LostConnection": "Strata pripojenia k serveru",
-"Nitrox_PlayerKicked": "Boli ste vyhodený zo serveru",
-// In game logging
-"Nitrox_UnableToConnect": "Nedá sa kontaktovať server na:",
-"Nitrox_StartServer": "Najprv spustite server aby ste sa mohli pripojiť do vášho sveta",
-"Nitrox_FirewallInterfering": "Vyzerá, že vám bránia vaše nastavenia firewallu",
-"Nitrox_InvalidUserName": "Prosím zadajte platné meno !\n\n Nemôžu v ňom byť medzery ani pochybné znaky.\n Povolené znaky : A-Z a-z 0-9 _ . -\nDĺžka : [3, 25]",
-"Nitrox_RequestingSessionPolicy": "Pýtam pravidlá relácie...",
-"Nitrox_RejectedSessionPolicy": "Rezervácia odmietnutá...",
-"Nitrox_DisconnectedSession": "Odpojený zo serveru",
-"Nitrox_Waiting": "Čakanie vo fronte pripojení...",
-"Nitrox_WaitingPassword": "Čakám na zadanie hesla...",
-"Nitrox_WaitingUserInput": "Čakám na vstup od používateľa...",
-"Nitrox_LaunchGame": "Spúštam hru...",
-"Nitrox_SilencedChatNotif": "Chat je teraz stlmený"
+    "Nitrox_Multiplayer": "Hrať spoločne",
+    "Nitrox_AddServer": "Pridať server",
+    "Nitrox_ConnectTo": "Pripojiť sa na",
+    "Nitrox_AddServerName": "Názov:",
+    "Nitrox_AddServerHost": "Adresa:",
+    "Nitrox_AddServerPort": "Port:",
+    "Nitrox_AddServerAdd": "Pridať",
+    "Nitrox_JoinServerPasswordHeader": "Server požaduje heslo",
+    "Nitrox_JoinServerIpAddress": "IP adresa servera",
+    "Nitrox_JoinServerPassword": "Heslo:",
+    "Nitrox_JoinServerFailed": "Nedá sa pripojiť na server",
+    "Nitrox_Cancel": "Zrušiť",
+    "Nitrox_Join": "Pripojiť sa",
+    "Nitrox_PlayerColor": "Farba hráča",
+    "Nitrox_EnterName": "Meno hráča",
+    "Nitrox_SubmitPassword": "Použiť",
+    "Nitrox_MultiplayerLoaded": "Multiplayer klient načítaný…",
+    "Nitrox_StreamerSettings": "Režim streamer",
+    "Nitrox_SilenceChat": "Vypnutie chatu",
+    "Nitrox_HideIp": "Skryť IP adresy",
+    "Nitrox_LoadingScreenWarn": "Očakávajte chyby, ktoré kazia hru",
+    "Nitrox_WorldSettling": "Čakám na nastavenia sveta",
+    "Nitrox_JoiningSession": "Pripájam sa do hry",
+    "Nitrox_SyncingWorld": "Synchronizujem svet",
+    "Nitrox_LostConnection": "Strata pripojenia k serveru",
+    "Nitrox_PlayerKicked": "Boli ste vyhodený zo serveru",
+    "Nitrox_UnableToConnect": "Nedá sa kontaktovať server na:",
+    "Nitrox_StartServer": "Najprv spustite server aby ste sa mohli pripojiť do vášho sveta",
+    "Nitrox_FirewallInterfering": "Vyzerá, že vám bránia vaše nastavenia firewallu",
+    "Nitrox_InvalidUserName": "Prosím zadajte platné meno !\n\n Nemôžu v ňom byť medzery ani pochybné znaky.\n Povolené znaky : A-Z a-z 0-9 _ . -\nDĺžka : [3, 25]",
+    "Nitrox_RequestingSessionPolicy": "Pýtam pravidlá relácie…",
+    "Nitrox_RejectedSessionPolicy": "Rezervácia odmietnutá…",
+    "Nitrox_DisconnectedSession": "Odpojený zo serveru",
+    "Nitrox_Waiting": "Čakanie vo fronte pripojení…",
+    "Nitrox_WaitingPassword": "Čakám na zadanie hesla…",
+    "Nitrox_WaitingUserInput": "Čakám na vstup od používateľa…",
+    "Nitrox_LaunchGame": "Spúštam hru…",
+    "Nitrox_SilencedChatNotif": "Chat je teraz stlmený"
 }
