+++ Nitrox.Assets.Subnautica/LanguageFiles/Spanish.json
@@ -1,45 +1,41 @@
 ﻿{
-"version": 4,
-// Main menu
-"Nitrox_Multiplayer": "Multijugador",
-"Nitrox_AddServer": "Añadir servidor",
-"Nitrox_ConnectTo": "Conectarse a",
-"Nitrox_AddServerName": "Nombre:",
-"Nitrox_AddServerHost": "Dirección:",
-"Nitrox_AddServerPort": "Port:",
-"Nitrox_AddServerAdd": "Añadir",
-"Nitrox_JoinServerPasswordHeader": "Contraseña requerido",
-"Nitrox_JoinServerIpAddress": "Dirección del servidor",
-"Nitrox_JoinServerPassword": "Contraseña:",
-"Nitrox_JoinServerFailed": "Error al conectar con el servidor",
-"Nitrox_Cancel": "Cancelar",
-"Nitrox_Join": "Entrar",
-"Nitrox_PlayerColor": "Color",
-"Nitrox_EnterName": "Nombre del jugador",
-"Nitrox_SubmitPassword": "Enviar contraseña",
-"Nitrox_MultiplayerLoaded": "Cliente multijugador cargado...",
-"Nitrox_StreamerSettings": "Modo streamer",
-"Nitrox_SilenceChat": "Silenciar el chat",
-"Nitrox_HideIp": "Ocultar IPs",
-// In game menu
-"Nitrox_LoadingScreenWarn": "Espere alguna inestabilidad",
-"Nitrox_WorldSettling": "Cargando mundo",
-"Nitrox_JoiningSession": "Conectando con el mundo multijugador",
-"Nitrox_SyncingWorld": "Sincronización del mundo multijugador",
-"Nitrox_LostConnection": "Conexión perdida al servidor",
-"Nitrox_PlayerKicked": "Expulsado del servidor",
-// In game logging
-"Nitrox_UnableToConnect": "Error al conectar con el servidor :",
-"Nitrox_StartServer": "Inicie su servidor para alcanzar a su mundo",
-"Nitrox_FirewallInterfering": "Parece que la configuración de tu cortafuegos está interfiriendo",
-"Nitrox_InvalidUserName": "Por favor, introduzca un nombre de jugador válido !\n\n No debe contener ningún espacio o caracteres dudosos\n Caracteres permitidos : A-Z a-z 0-9 _ . -\nLargo : [3, 25]",
-"Nitrox_RequestingSessionPolicy": "Conectando con el servidor...",
-"Nitrox_RejectedSessionPolicy": "Reservación rechazada....",
-"Nitrox_DisconnectedSession": "Desconectado del servidor",
-"Nitrox_Waiting": "Esperando en cola para unirse...",
-"Nitrox_WaitingPassword": "Esperando contraseña...",
-"Nitrox_WaitingUserInput": "Esperando información del jugador...",
-"Nitrox_LaunchGame": "Lanzamiento de la partida...",
-"Nitrox_SilencedChatNotif": "Se silencia el chat",
-"Nitrox_CommandNotAvailable": "Este comando no está disponible con Nitrox"
+    "Nitrox_Multiplayer": "Multijugador",
+    "Nitrox_AddServer": "Añadir servidor",
+    "Nitrox_ConnectTo": "Conectarse a",
+    "Nitrox_AddServerName": "Nombre:",
+    "Nitrox_AddServerHost": "Dirección:",
+    "Nitrox_AddServerPort": "Port:",
+    "Nitrox_AddServerAdd": "Añadir",
+    "Nitrox_JoinServerPasswordHeader": "Contraseña requerido",
+    "Nitrox_JoinServerIpAddress": "Dirección del servidor",
+    "Nitrox_JoinServerPassword": "Contraseña:",
+    "Nitrox_JoinServerFailed": "Error al conectar con el servidor",
+    "Nitrox_Cancel": "Cancelar",
+    "Nitrox_Join": "Entrar",
+    "Nitrox_PlayerColor": "Color",
+    "Nitrox_EnterName": "Nombre del jugador",
+    "Nitrox_SubmitPassword": "Enviar contraseña",
+    "Nitrox_MultiplayerLoaded": "Cliente multijugador cargado…",
+    "Nitrox_StreamerSettings": "Modo streamer",
+    "Nitrox_SilenceChat": "Silenciar el chat",
+    "Nitrox_HideIp": "Ocultar IPs",
+    "Nitrox_LoadingScreenWarn": "Espere alguna inestabilidad",
+    "Nitrox_WorldSettling": "Cargando mundo",
+    "Nitrox_JoiningSession": "Conectando con el mundo multijugador",
+    "Nitrox_SyncingWorld": "Sincronización del mundo multijugador",
+    "Nitrox_LostConnection": "Conexión perdida al servidor",
+    "Nitrox_PlayerKicked": "Expulsado del servidor",
+    "Nitrox_UnableToConnect": "Error al conectar con el servidor :",
+    "Nitrox_StartServer": "Inicie su servidor para alcanzar a su mundo",
+    "Nitrox_FirewallInterfering": "Parece que la configuración de tu cortafuegos está interfiriendo",
+    "Nitrox_InvalidUserName": "Por favor, introduzca un nombre de jugador válido !\n\n No debe contener ningún espacio o caracteres dudosos\n Caracteres permitidos : A-Z a-z 0-9 _ . -\nLargo : [3, 25]",
+    "Nitrox_RequestingSessionPolicy": "Conectando con el servidor…",
+    "Nitrox_RejectedSessionPolicy": "Reservación rechazada…",
+    "Nitrox_DisconnectedSession": "Desconectado del servidor",
+    "Nitrox_Waiting": "Esperando en cola para unirse…",
+    "Nitrox_WaitingPassword": "Esperando contraseña…",
+    "Nitrox_WaitingUserInput": "Esperando información del jugador…",
+    "Nitrox_LaunchGame": "Lanzamiento de la partida…",
+    "Nitrox_SilencedChatNotif": "Se silencia el chat",
+    "Nitrox_CommandNotAvailable": "Este comando no está disponible con Nitrox"
 }
