+++ Nitrox.Assets.Subnautica/LanguageFiles/Turkish.json
@@ -1,44 +1,40 @@
 {
-"version": 3,
-// Main menu
-"Nitrox_Multiplayer": "Multiplayer",
-"Nitrox_AddServer": "Sunucu Ekle",
-"Nitrox_ConnectTo": "Sunucuya Bağlan",
-"Nitrox_AddServerName": "İsim:",
-"Nitrox_AddServerHost": "Sunucu:",
-"Nitrox_AddServerPort": "Port:",
-"Nitrox_AddServerAdd": "Ekle",
-"Nitrox_JoinServerPasswordHeader": "Sunucu Şifresi Gerekli",
-"Nitrox_JoinServerIpAddress": "Sunucu IP",
-"Nitrox_JoinServerPassword": "Şifre:",
-"Nitrox_JoinServerFailed": "Sunucuya bağlanılamadı",
-"Nitrox_Cancel": "İptal",
-"Nitrox_Join": "Katıl",
-"Nitrox_PlayerColor": "Oyuncu Rengi",
-"Nitrox_EnterName": "Oyuncu İsmi",
-"Nitrox_SubmitPassword": "Gönder",
-"Nitrox_MultiplayerLoaded": "Multiplayer istemcisi yüklendi...",
-"Nitrox_StreamerSettings": "Streamer modu",
-"Nitrox_SilenceChat": "Sohbeti sessize al",
-"Nitrox_HideIp": "IP adreslerini gizle",
-// In game menu
-"Nitrox_LoadingScreenWarn": "Oyunu bozacak hatalar",
-"Nitrox_WorldSettling": "Dünya Yerleşimi Bekleniyor",
-"Nitrox_JoiningSession": "Sunucuya Bağlanılıyor",
-"Nitrox_SyncingWorld": "Dünya Seknronize Ediliyor",
-"Nitrox_LostConnection": "Sunucu bağlantısı kaybedildi",
-"Nitrox_PlayerKicked": "Sunucudan atıldınız",
-// In game logging
-"Nitrox_UnableToConnect": "Bu uzak sunucu ile iletişim kurulamıyor:",
-"Nitrox_StartServer": "Kendi dünyanıza katılmak için önce sunucu başlatmalısınız",
-"Nitrox_FirewallInterfering": "Güvenlik duvarı ayarları düzgün yapılmamış görünüyor",
-"Nitrox_InvalidUserName": "Lütfen geçerli bir kullanıcı adı girin!\n\n Boşluk ve özel karakter içermemelidir\n İzin verilen karakterler: A-Z a-z 0-9 _ . -\nUzunluk : [3, 25]",
-"Nitrox_RequestingSessionPolicy": "Oturum bilgileri alınıyor...",
-"Nitrox_RejectedSessionPolicy": "Rezervasyon reddedildi...",
-"Nitrox_DisconnectedSession": "Sunucu bağlantısı kesildi",
-"Nitrox_Waiting": "Bağlantı kuyruğunda beklemek...",
-"Nitrox_WaitingPassword": "Sunucu şifresi girilmesi bekleniyor...",
-"Nitrox_WaitingUserInput": "Kullanıcı adı girilmesi bekleniyor...",
-"Nitrox_LaunchGame": "Oyun başlatılıyor...",
-"Nitrox_SilencedChatNotif": "Sohbet artık sessiz"
+    "Nitrox_Multiplayer": "Multiplayer",
+    "Nitrox_AddServer": "Sunucu Ekle",
+    "Nitrox_ConnectTo": "Sunucuya Bağlan",
+    "Nitrox_AddServerName": "İsim:",
+    "Nitrox_AddServerHost": "Sunucu:",
+    "Nitrox_AddServerPort": "Port:",
+    "Nitrox_AddServerAdd": "Ekle",
+    "Nitrox_JoinServerPasswordHeader": "Sunucu Şifresi Gerekli",
+    "Nitrox_JoinServerIpAddress": "Sunucu IP",
+    "Nitrox_JoinServerPassword": "Şifre:",
+    "Nitrox_JoinServerFailed": "Sunucuya bağlanılamadı",
+    "Nitrox_Cancel": "İptal",
+    "Nitrox_Join": "Katıl",
+    "Nitrox_PlayerColor": "Oyuncu Rengi",
+    "Nitrox_EnterName": "Oyuncu İsmi",
+    "Nitrox_SubmitPassword": "Gönder",
+    "Nitrox_MultiplayerLoaded": "Multiplayer istemcisi yüklendi…",
+    "Nitrox_StreamerSettings": "Streamer modu",
+    "Nitrox_SilenceChat": "Sohbeti sessize al",
+    "Nitrox_HideIp": "IP adreslerini gizle",
+    "Nitrox_LoadingScreenWarn": "Oyunu bozacak hatalar",
+    "Nitrox_WorldSettling": "Dünya Yerleşimi Bekleniyor",
+    "Nitrox_JoiningSession": "Sunucuya Bağlanılıyor",
+    "Nitrox_SyncingWorld": "Dünya Seknronize Ediliyor",
+    "Nitrox_LostConnection": "Sunucu bağlantısı kaybedildi",
+    "Nitrox_PlayerKicked": "Sunucudan atıldınız",
+    "Nitrox_UnableToConnect": "Bu uzak sunucu ile iletişim kurulamıyor:",
+    "Nitrox_StartServer": "Kendi dünyanıza katılmak için önce sunucu başlatmalısınız",
+    "Nitrox_FirewallInterfering": "Güvenlik duvarı ayarları düzgün yapılmamış görünüyor",
+    "Nitrox_InvalidUserName": "Lütfen geçerli bir kullanıcı adı girin!\n\n Boşluk ve özel karakter içermemelidir\n İzin verilen karakterler: A-Z a-z 0-9 _ . -\nUzunluk : [3, 25]",
+    "Nitrox_RequestingSessionPolicy": "Oturum bilgileri alınıyor…",
+    "Nitrox_RejectedSessionPolicy": "Rezervasyon reddedildi…",
+    "Nitrox_DisconnectedSession": "Sunucu bağlantısı kesildi",
+    "Nitrox_Waiting": "Bağlantı kuyruğunda beklemek…",
+    "Nitrox_WaitingPassword": "Sunucu şifresi girilmesi bekleniyor…",
+    "Nitrox_WaitingUserInput": "Kullanıcı adı girilmesi bekleniyor…",
+    "Nitrox_LaunchGame": "Oyun başlatılıyor…",
+    "Nitrox_SilencedChatNotif": "Sohbet artık sessiz"
 }
