+++ NitroxPatcher/Patches/Persistent/Language_LoadLanguageFile_Patch.cs
@@ -6,55 +6,54 @@ using HarmonyLib;
 using LitJson;
 using NitroxModel.Helper;
 
-namespace NitroxPatcher.Patches.Persistent
+namespace NitroxPatcher.Patches.Persistent;
+
+public class Language_LoadLanguageFile_Patch : NitroxPatch, IPersistentPatch
 {
-    internal class Language_LoadLanguageFile_Patch : NitroxPatch, IPersistentPatch
+    private static readonly MethodInfo TARGET_METHOD = Reflect.Method((Language t) => t.LoadLanguageFile(default));
+
+    public static void Postfix(string language, Dictionary<string, string> ___strings)
     {
-        private static readonly MethodInfo TARGET_METHOD = Reflect.Method((Language t) => t.LoadLanguageFile(default(string)));
+        string[] files = {
+            Path.Combine(NitroxUser.LauncherPath, "LanguageFiles", "English.json"), // Using English as fallback.
+            Path.Combine(NitroxUser.LauncherPath, "LanguageFiles", $"{language}.json")
+        };
 
-        public static void Postfix(string language, Dictionary<string, string> ___strings)
+        foreach (string file in files)
         {
-            string[] files = {
-                Path.Combine(NitroxUser.LauncherPath, "LanguageFiles", "English.json"), // Using English as fallback.
-                Path.Combine(NitroxUser.LauncherPath, "LanguageFiles", $"{language}.json")
-            };
-
-            foreach (string file in files)
+            if (!File.Exists(file))
             {
-                if (!File.Exists(file))
-                {
-                    Log.Warn($"No language file was found for at {file}. Using English as fallback");
-                    continue;
-                }
+                Log.Warn($"No language file was found for at {file}. Using English as fallback");
+                continue;
+            }
 
-                JsonData json;
-                using (StreamReader streamReader = new(file))
+            JsonData json;
+            using (StreamReader streamReader = new(file))
+            {
+                try
                 {
-                    try
-                    {
-                        json = JsonMapper.ToObject(streamReader);
-                    }
-                    catch (Exception ex)
-                    {
-                        Log.Error(ex, $"Error while reading language file {language}.json");
-                        return;
-                    }
+                    json = JsonMapper.ToObject(streamReader);
                 }
-
-                foreach (string key in json.Keys)
+                catch (Exception ex)
                 {
-                    if (json[key].IsString)
-                    {
-                        ___strings[key] = (string)json[key];
-                    }
+                    Log.Error(ex, $"Error while reading language file {language}.json");
+                    return;
                 }
             }
 
+            foreach (string key in json.Keys)
+            {
+                if (json[key].IsString)
+                {
+                    ___strings[key] = (string)json[key];
+                }
+            }
         }
 
-        public override void Patch(Harmony harmony)
-        {
-            PatchPostfix(harmony, TARGET_METHOD);
-        }
+    }
+
+    public override void Patch(Harmony harmony)
+    {
+        PatchPostfix(harmony, TARGET_METHOD);
     }
 }
